% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInteraction.R
\name{addInteraction}
\alias{addInteraction}
\title{Set up interactions among covariates}
\usage{
addInteraction(.data, which_covs = "pairs", exclude = NULL)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{which_covs}{List with vector(s) of names of covariates to interact, e.g. list(c('A', 'C', 'D')), or 'pairs' for all pairwise interactions.}

\item{exclude}{For option 'pairs', exclude the covariates named from interactions.}
}
\description{
The module constructs new covariates torepresent interactions (and possibly polynomial terms).
}
\details{
Statistical interactions are (mathematically) the product of covariates. This module allows the user to pre-specify which interactions to add to the data. If a four-way interaction between A, B, C and D is desired, it automatically adds all possible two- and three-way interactions as well (obeying the marginality theorem). If the same variable name is provided several times in a vector (e.g. c(A, A, A)), this will lead to a cubic polynomial.
}
\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\section{Version}{
 0
}

\section{Date submitted}{
  2016-06-15
}
\author{
Alison Johnston & Carsten F. Dormann, \email{alison.johnston@bto.org}
}
\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}
}

