% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/AirNCEP.R
\name{AirNCEP}
\alias{AirNCEP}
\title{Covariate module: AirNCEP}
\usage{
AirNCEP()
}
\description{
Covariate module to grab a coarse resolution mean air temperature raster from
 January-February 2001-2002 for the UK.
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
\code{\link{RNCEP::NCEP.gather}}

Other covariate: \code{\link{Bioclim}};
  \code{\link{LocalRaster}}; \code{\link{NCEP}};
  \code{\link{UKAir}}; \code{\link{UKBioclim}}
}

