% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesPlumbeus.R
\name{AnophelesPlumbeus}
\alias{AnophelesPlumbeus}
\title{Occurrence module: AnophelesPlumbeus}
\usage{
AnophelesPlumbeus(extent = c(-10, 10, 45, 65))
}
\arguments{
\item{extent}{A numeric vector of length 4 giving the coordinates of the 
rectangular region within which to carry out the analysis, in the 
order: xmin, xmax, ymin, ymax. By default the extent of the UK.}
}
\description{
Occurrence module to grab *Anopheles plumbeus* (a mosquito) presence
      (presence-only) data from GBIF, in the area bounded by extent.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence-only
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other occurrence: \code{\link{CWBZimbabwe}},
  \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{UKAnophelesPlumbeus}}
}

