% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appify.R
\name{Appify}
\alias{Appify}
\title{Appify}
\usage{
Appify(.model, .ras, directory = tempdir(), run_App = FALSE)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.#'}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.#'}

\item{directory}{(Optional) a string giving the path to a directory within which to save the shiny app. This directory must exist.#'}

\item{run_App}{Logical, should the app be run automatically. Not recommended for workflows containing lists. Default is FALSE}
}
\description{
The Appify module builds a shiny app that displays dynamic graphical and tabular representations of the input data (occurrence and covarite) as well as model predictions.
}
\details{
Appify will return the path to the shiny application. For details on how to share the application visit http://shiny.rstudio.com/tutorial/lesson7/. You can run the app locally using shiny::runApp(appDir = <YOUR_PATH>)
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2016-05-05
}

\section{Data type}{
 presence-only, presence/absence, presence/background, abundance, proportion
}
\author{
ZOON Developers, Tom August, \email{zoonproject@gmail.com}
}
\seealso{
Other output: \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}}, \code{\link{NoOutput}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}

