% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Background.R
\name{Background}
\alias{Background}
\title{Process module: Background}
\usage{
Background(.data, n = 100)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{n}{the number of background points to sample}
}
\description{
Process module to generate background records at random in
     cells of the covariate raster and return these along with the presence only data.
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other process: \code{\link{BackgroundAndCrossvalid}};
  \code{\link{BiasBackground}};
  \code{\link{Crossvalidate}}; \code{\link{NoProcess}};
  \code{\link{OneHundredBackground}};
  \code{\link{OneThousandBackground}};
  \code{\link{PartitionDisc}}
}

