% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BackgroundAndCrossvalid}
\alias{BackgroundAndCrossvalid}
\title{Process module: BackgroundAndCrossvalid}
\usage{
BackgroundAndCrossvalid(occurrence, ras, k = 5)
}
\description{
Process module to generate up to 100 background records at random in
     cells of ras and split all data in k folds for cross validation.
}

