% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BiasBackground.R
\name{BiasBackground}
\alias{BiasBackground}
\title{BiasBackground: Generate background data using a bias grid}
\usage{
BiasBackground(.data, bias, n = 100)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}.
\code{.data} is a list of a data frame and a raster object returned from
occurrence modules and covariate modules respectively. \code{.data} is a list
of a data frame and a raster object returned from occurrence modules and
covariate modules respectively. \code{.data} is passed automatically in
workflow from the occurrence and covariate modules to the process module(s)
and should not be passed by the user.}

\item{bias}{Raster with cells giving the relative probability of a background record being sampled there.}

\item{n}{Positive integer giving the number of background points required.}
}
\description{
This process module genereates a sample of background records according to some spatial model of recording bias,indicated by a raster layer.
}
\details{
Module type: Process
}
\author{
Nick Golding
}

