% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioclim_future.R
\name{Bioclim_future}
\alias{Bioclim_future}
\title{Covariate module: Bioclim_future}
\usage{
Bioclim_future(extent = c(-10, 10, 45, 65), resolution = 10,
  layers = 1:19, rcp = 45, model = "AC", year = 70)
}
\arguments{
\item{extent}{Length 4 numeric vector giving min longitude, max longitude, min latitude, max latitude.}

\item{resolution}{Resolution in minutes. Must be one of 2.5, 5 or 10. Default is 10.}

\item{layers}{which bioclim layers to obtain, a vector of integers between 1 and 19. See \url{http://www.worldclim.org/bioclim}.}

\item{rcp}{numeric. Representative Concentration Pathways, or greenhouse gases concentration trajectories (see \url{https://en.wikipedia.org/wiki/Representative_Concentration_Pathways}. Should be one of 26, 45, 60, or 85.}

\item{model}{character. General Circulation Model. Should be one of "AC", "BC", "CC", "CE", "CN", "GF", "GD", "GS", "HD", "HG", "HE", "IN", "IP", "MI", "MR", "MC", "MP", "MG", or "NO".}

\item{year}{numeric. Time period. Should be either 50 (which represents the period 2041-2060) or 70 (which represents the period 2061-2080).}
}
\description{
Get worldclim data for the future (CMIP5). Downloads then stores locally.
}
\section{Version}{
 1.0.0
}

\section{Date submitted}{
 2016-06-15
}

\seealso{
\url{http://www.worldclim.org} and \code{\link[raster]{getData}}.

Other covariate: \code{\link{AirNCEP}},
  \code{\link{Bioclim}}, \code{\link{CarolinaWrenRasters}},
  \code{\link{LocalRaster}}, \code{\link{NCEP}},
  \code{\link{NaiveRandomRaster}}, \code{\link{UKAir}},
  \code{\link{UKBioclim}}
}
\author{
F. Rodriguez-Sanchez & E. Van Loon, \email{zoonproject@gmail.com}
}
