% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BioclimNewArea}
\alias{BioclimNewArea}
\title{Covariate module: BioclimNewArea}
\usage{
BioclimNewArea(trainExtent, predictExtent, resolution = 5)
}
\arguments{
\item{trainExtent}{Length 4 character vector giving min lat, max lat, min long, max long.}

\item{predictExtent}{Length 4 character vector giving min lat, max lat, min long, max long.}

\item{resolution}{Resolution in minutes. Must be one of 2.5, 5 or 10. Default is 5.}
}
\description{
Get worldclim environment data. Downloads then stores locally.
}

