% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BiomodModel}
\alias{BiomodModel}
\title{Model module wrapper for BIOMOD_Modeling()}
\arguments{
\item{modelType}{A character vector to describe models to use. Select from
'GLM','GBM','GAM','CTA','ANN','SRE','FDA','MARS','RF','MAXENT'}
}
\value{
a Raster* object (class from the raster package) with the gridded
     covariates used to train and predict from the SDM.
}
\description{
Model module wrapper for BIOMOD_Modeling()
}
\seealso{
\code{\link{biomod2::BIOMOD_ModelingOptions}}
}

