% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{Bootstrap}
\alias{Bootstrap}
\title{Process module: Boostrap}
\usage{
Bootstrap(.data, n = NULL, seed = NULL)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow
function}. \code{.data} is a list of a data frame and a raster object
returned from occurrence modules and covariate modules respectively.
\code{.data} is passed automatically in workflow from the occurrence and
covariate modules to the process module(s) and should not be passed by the
user.}

\item{n}{the number of points to sample. Defaults to the same number as the dataset}

\item{seed}{Numeric used with \code{\link[base]{set.seed}}}
}
\description{
Process module to generate random bootstraps of the 
  data.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2016-03-23
}

\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}},
  \code{\link{LonLatToCovariates}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}}, \code{\link{RemoveNAs}},
  \code{\link{StandardiseCov}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}},
  \code{\link{spThin}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
