% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarolinaWrenRasters.R
\name{CarolinaWrenRasters}
\alias{CarolinaWrenRasters}
\title{Covariate rasters for the USA used in the Carolina Wren analysis}
\usage{
CarolinaWrenRasters()
}
\description{
Land cover and coordinate covariate data for the USA, used in the Carolina Wren MaxLike analysis. These data are made available in the maxlike R package. See \code{\link[maxlike]{carw.data}} for more details, and see the modules \code{CarolinaWrenPA} and \code{CarolinaWrenPO} for corresponding occurrence data. This module loads the data, converts them into a \code{RasterStack} object and transforms the coordinates to lat/longs.
}
\details{
The six covariates are: Percentage cover of mixed forest, deciduous forest, coniferous forest and grassland, and the latitude and longitude of each cell. After converting to latlongs, the pixels have dimensions 0.29x0.223 decimal degrees and the extent is \code{c(-138.71, -52.58, 18.15, 54.95)}
}
\section{Version}{
 0
}

\section{Date submitted}{
  2016-05-21
}

\seealso{
Other covariate: \code{\link{AirNCEP}},
  \code{\link{Bioclim_future}}, \code{\link{Bioclim}},
  \code{\link{LocalRaster}}, \code{\link{NCEP}},
  \code{\link{NaiveRandomRaster}}, \code{\link{UKAir}},
  \code{\link{UKBioclim}}
}
\author{
Nick Golding, \email{nick.golding.research@gmail.com}
}
