% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarolinaWrenValidation.R
\name{CarolinaWrenValidation}
\alias{CarolinaWrenValidation}
\title{Presence/absence validation data for the Carolina Wren in the USA}
\usage{
CarolinaWrenValidation(.data)
}
\description{
Presence/absence of the Carolina Wren \emph{Thryothorus 
  ludovicianus} at Breeding Bird Survey transects in 2006. These data are 
  made available in the maxlike R package. See 
  \code{\link[maxlike]{carw.data}} for more details. This module loads the 
  data, transforms the coordinates to lat/longs and formats the data for use 
  in a zoon workflow. This module is identical to \code{CarolinaWrenPO}, 
  except that the data are added as external validation data (fold = 0),
  rather than training data (fold = 1).
}
\details{

}
\section{Data type}{
 presence/absence
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2016-12-21
}
\author{
Nick Golding, \email{nick.golding.research@gmail.com}
}
\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}

