% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{Clean}
\alias{Clean}
\title{Process module: Clean}
\usage{
Clean(.data, which = c(1, 2, 3, 4))
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{which}{A numeric or character vector describing which data cleaning processes to use (see details).}
}
\description{
Do some data cleaning on occurrence points
}
\details{
A number of data cleaning steps are implemented. \code{which} takes a (case insensitive) character or numeric vector
 declaring which should be used:
\itemize{
 \item{"1 - Impossible"}{Remove rows with imposisble latitude or longitude values}
 \item{"2 - Incomplete"}{Remove rows with missing data for either latitude of longitude}
 \item{"3 - Unlikely"}{Remove rows with unlikely data such as (0, 0)}
}
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-03-21
}

\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
\code{\link{scrubr::coord_incomplete}}

Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}

