% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefficientPlot.R
\name{CoefficientPlot}
\alias{CoefficientPlot}
\title{Plot estimated coefficients from LogisticRegression}
\usage{
CoefficientPlot(.model, .ras)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}
}
\description{
This module plots the estimated coefficients from the LogisticRegression module and returns the summarised model object
}
\details{
The module outputs a point plot of coefficient estimates from the LogisticRegression module. Coefficients with a p-value <= 0.05 are identified on the graph. The summary glm object is returned in the console. This module is only implemented for the LogisticRegression model module.
}
\section{Data type}{
 presence/absence
}

\section{Version}{
 0.3
}

\section{Date submitted}{
  2017-05-09
}

\seealso{
Other output: \code{\link{AUC}}, \code{\link{Appify}},
  \code{\link{CovHistograms}}, \code{\link{DataSummary}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{InteractiveOccurrenceMap}},
  \code{\link{NoOutput}}, \code{\link{PairPlot}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ReliabilityPlot}},
  \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}}, \code{\link{ResponsePlot}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
Liz Martin, \email{emartin@student.unimelb.edu.au}
}
