% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWBZimbabwe.R
\name{CWBZimbabwe}
\alias{CWBZimbabwe}
\title{Occurrence module: coffee white Stem borer surveys in Zimbabwe}
\usage{
CWBZimbabwe()
}
\description{
This is an example presence/absence dataset for the
coffee white stem borer \emph{Monochamus leuconotus} P. taken from an
open access dataset on the Dryad data repository.
The data are made available by those authors under a Creative Commons CC0
These data are used in the module CWBZimbabwe which can be used for running
toy presence/absence species distribution models.
}
\details{
Original publication:
Kutywayo D, Chemura A, Kusena W, Chidoko P, Mahoya C (2013) The impact of
 climate change on the potential distribution of agricultural pests: the case
  of the coffee white stem borer (\emph{Monochamus leuconotus} P.) in
   Zimbabwe. PLoS ONE 8(8): e73432.

Dryad data package:
   Kutywayo D, Chemura A, Kusena W, Chidoko P, Mahoya C (2013) Data from:
   The impact of climate change on the potential distribution of agricultural
    pests: the case of the coffee white stem borer (\emph{Monochamus
     leuconotus} P.) in Zimbabwe. Dryad Digital Repository.
  \url{http://dx.doi.org/10.1371/journal.pone.0073432}
}

