% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preview.R
\name{datapreview}
\alias{datapreview}
\title{Data Preview Server}
\usage{
datapreview(input, output, session, data, n = reactive(100))
}
\arguments{
\item{data}{data.frame that you want to preview}

\item{n}{number of observations to show. (Default: 100)}
}
\description{
Data Preview Server
}
\seealso{
\code{\link{datapreviewUI}}
}

