% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preview.R
\name{datapreviewUI}
\alias{datapreviewUI}
\title{Data Preview UI}
\usage{
datapreviewUI(id, title = "Preview", width = 12)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{title}{Title of the box (Default: "Preview")}

\item{width}{Wdth of the shinydashboard box. (Default: 12)}
}
\description{
Data Preview UI
}

