% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{DataSummary}
\alias{DataSummary}
\title{Output module: DataSummary}
\usage{
DataSummary(.model, .ras)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}
}
\description{
The DataSummary module generates summaries of both the point data (non-background) and the raster data for the purposes of data exploration. The module outputs a list of two summaries to the console.
}
\details{

}
\section{Data type}{
 presence-only, presence/background, presence/absence
}

\section{Version}{
 1.1
}

\section{Date submitted}{
  2017-10-05
}

\seealso{
Other output: \code{\link{AIC}}, \code{\link{AUC}},
  \code{\link{Appify}}, \code{\link{CoefficientPlot}},
  \code{\link{CovHistograms}}, \code{\link{Deviance}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{InteractiveOccurrenceMap}},
  \code{\link{NoOutput}}, \code{\link{PairPlot}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ReliabilityPlot}},
  \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}}, \code{\link{ResponsePlot}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
David Wilkinson, \email{davidpw@student.unimelb.edu.au}
}
\concept{output}
