% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-select.R
\name{dynamicselect}
\alias{dynamicselect}
\title{Dynamic Select Server}
\usage{
dynamicselect(input, output, session, label, choices)
}
\arguments{
\item{label}{label text in the input}

\item{choices}{vector with possible values}
}
\description{
Dynamic Select Server
}
\seealso{
\code{\link{dynamicselectUI}}
}

