% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-select.R
\name{dynamicselectUI}
\alias{dynamicselectUI}
\title{Dynamic Selectize Input 
Dynamic Select UI}
\usage{
dynamicselectUI(id)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}
}
\description{
Dynamic Selectize Input 
Dynamic Select UI
}
\seealso{
\code{\link{dynamicselect}}
}

