% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Figshare.R
\name{Figshare}
\alias{Figshare}
\title{Output module: Figshare}
\usage{
Figshare(model, ras, title, description = "zoon workflow", authors = "zoon",
  categories = "SDM", tags = "zoon")
}
\arguments{
\item{title}{String giving the title}

\item{description}{String describing the workflow}

\item{authors}{Charactor vector of full authors names}

\item{categories}{Character vector of figshare categories e.g. ecology.}

\item{tags}{Character vector of searchable tags.}
}
\description{
Upload your workflow call, model, occurrance and covariate data to figshare as a RData file. Uploads privately but be aware I cannot assure the privacy of any data. This module is a work in progress, and will obviously want to upload the output from the output modules.
}

