% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\alias{use}
\title{Import objects into module}
\usage{
import(from, ..., into = parent.env(parent.frame()))

use(module, into = parent.env(parent.frame()))
}
\arguments{
\item{from}{a package name}

\item{...}{names to import from package}

\item{into}{the environment where the thing is stored. By default the parent
of the module}

\item{module}{a module, aka a list, or any named collection}
}
\description{
Import objects into module
}

