% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/InteractiveMap.R
\name{InteractiveMap}
\alias{InteractiveMap}
\title{Output module: InteractiveMap}
\usage{
InteractiveMap(model, ras)
}
\arguments{
\item{model}{}

\item{ras}{}
}
\description{
Plot an interactive map of the predicted distribution.
}

