% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveOccurrenceMap.R
\name{InteractiveOccurrenceMap}
\alias{InteractiveOccurrenceMap}
\title{Interactive Occurrence Map}
\usage{
InteractiveOccurrenceMap(.model, .ras)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}
}
\description{
This output module creates an interactive occurrence map of the species data.
}
\details{
The module creates an html widget that displays in  the Viewer panel
}
\section{Data type}{
 presence-only, presence/absence
}

\section{Version}{
 0.2
}

\section{Date submitted}{
  2017-06-08
}

\seealso{
Other output: \code{\link{AUC}}, \code{\link{Appify}},
  \code{\link{CoefficientPlot}},
  \code{\link{CovHistograms}}, \code{\link{DataSummary}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}}, \code{\link{NoOutput}},
  \code{\link{PairPlot}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}}, \code{\link{ResponsePlot}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
David Wilkinson, \email{davidpw@student.unimelb.edu.au}
}
