% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JitterOccurrence.R
\name{JitterOccurrence}
\alias{JitterOccurrence}
\title{Add noise to occurrence coordinates}
\usage{
JitterOccurrence(.data, sd = 0)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{sd}{standard deviation of the gaussian noise to add the the coordinates of each ocurrence record}
}
\description{
Often the geographic coordinates of occurrence records are subject to uncertainty. One approach ot accounting for that uncertainty in an SDM analysis is to run several analyses, each time sampling a different location from a probability distribution over the likely "true" coordinates. \code{JitterOccurrence} generates one such sample, under an assumption that the distribution is an isotropic Gaussian with standard deviation \code{sd}
}
\details{
Currently only a single standard deviation is provided for all occurrence records. Future iterations of this module will enable more control over the uncertainty distributions
}
\section{Data type}{
 presence-only, presence/absence, presence/background, abundance, proportion
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2016-06-16
}
\author{
Nick Golding, \email{nick.golding.research@gmail.com}
}
\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{MESSMask}}, \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}

