% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LASSO}
\alias{LASSO}
\title{Model module: LASSO}
\usage{
LASSO(df, alpha = 1)
}
\arguments{
\item{alpha}{Elasticnet mixing parameter. 1 (default) is LASSO. 0 is ridge regression.}
}
\description{
Model module to fit LASSO or ridge regression.
}
\seealso{
\code{\link{glmnet}}
}

