% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LocalData}
\alias{LocalData}
\title{Format local occurrence data to be used with zoon.}
\arguments{
\item{filename}{The path to the file. Currently assumed a .csv file.}

\item{occurrenceType}{What type data is it?
One of 'presence', 'presence/absence', 'abundance')}
}
\description{
Format local occurrence data to be used with zoon.
}
\seealso{
\code{\link{read.csv}}
}

