% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LocalRaster.R
\name{LocalRaster}
\alias{LocalRaster}
\title{LocalRaster: Read in a single raster or a list of rasters and stack them.}
\usage{
LocalRaster(filenames)
}
\arguments{
\item{filenames}{Either a string of the filename of the raster layer, or a list of strings of filenames to rasters that will be stacked.}
}
\description{
Read in a single raster or a list of rasters and stack them.
}
\details{
Module Type: Covariate
}
\author{
Tim Lucas

Email: timcdlucas@gmail.com
}

