% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{LogisticRegression}
\alias{LogisticRegression}
\title{Model module: LogisticRegression}
\usage{
LogisticRegression(df)
}
\description{
Model module to fit a simple logistic regression model
}
\seealso{
\code{\link{glm}}
}

