% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LonLatToCovariates.R
\name{LonLatToCovariates}
\alias{LonLatToCovariates}
\title{Create new covariates using latitude and longitude values}
\usage{
LonLatToCovariates(.data, lat = TRUE, lon = TRUE)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function.} \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. .data is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{lat}{Logical, return the latitude column?}

\item{lon}{Logical, return the longitude column?}
}
\description{
Create two new covariates, Longitudecov and Latitudecov that are simply the latitude and longitude for each cell. These covariates can be used in otherwise non-spatial models. Some specifically spatial models such as INLA require these covariates and will use the specifically named covariates once implemented.
}
\details{

}
\section{Data type}{
 presence-only, presence/absence, presence/background, abundance, proportion
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2017-04-25
}

\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}}, \code{\link{RemoveNAs}},
  \code{\link{StandardiseCov}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}},
  \code{\link{spThin}}
}
\author{
Tim Lucas, \email{timcdlucas@gmail.com}
}
\concept{process}
