% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorem_ipsum_UK.R
\name{Lorem_ipsum_UK}
\alias{Lorem_ipsum_UK}
\title{A dataset of Lorem ipsum occurrences}
\usage{
Lorem_ipsum_UK()
}
\description{
The module retrieves a dataset ofLorem ipsum records from figshare. This dataset containsprecence only data and was collected between 1990 and2000 by members of to Lorem ipsum appreciation society
}
\details{
This dataset is fake, Lorem ipsum does not exist
}
\section{Data type}{
 presence-only
}

\section{Version}{
 0
}

\section{Date submitted}{
  2016-06-13
}

\seealso{
Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceData}}, \code{\link{NATrees}},
  \code{\link{NBNdataByName}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{SugarMaple}},
  \code{\link{UKAnophelesPlumbeus}}
}
\author{
A.B. Ceidi, \email{ABCD@anemail.com}
}
