% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearn.R
\name{MachineLearn}
\alias{MachineLearn}
\title{Model module: MachineLearn}
\usage{
MachineLearn(.df, method = "glmnet", tuneLength = 8, metric = "ROC",
  number = 5, repeats = 1, tuneGrid = NULL, trControl = NULL, ...)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}.
\code{.df} is data frame that combines the occurrence}

\item{method}{The machine learning method to use. Common examples are "glm", 
"nnet", "gbm", "glmnet", "rf". See 
http://topepo.github.io/caret/modelList.html for a full list. Only 
classification or dual use models are useable.}

\item{tuneLength}{How many values of each tuning/hyperparameter should be tried?}

\item{metric}{a string that specifies what summary metric will be used to
select the optimal model. Options are "ROC", "Accuracy" and
"Kappa".}

\item{number}{How many folds to use in cross validation.}

\item{repeats}{How many times should the entire cross validation process be 
repeated. Increasing this will reduce instability in your model performace,
but will take longer to run.}

\item{tuneGrid}{Explicitely pass a data frame of tuning/hyperparameter combinations. If
NULL, tuneLength will be used instead.}

\item{trControl}{A named list of further arguments to pass to trainControl. See 
\code{\link[caret]{trainControl}} for details.}

\item{...}{Other arguments passed to \code{\link[caret]{train}}.}
}
\description{
Model module to fit a very large number of machine learning models.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence/absence
}
\author{
Tim CD Lucas, \email{timcdlucas@gmail.com}
}
\seealso{
\code{\link{caret::train}} \code{\link{trainControl}}

Other model: \code{\link{BiomodModel}}, \code{\link{GBM}},
  \code{\link{LogisticRegression}}, \code{\link{MaxEnt}},
  \code{\link{OptGRaF}}, \code{\link{QuickGRaF}},
  \code{\link{RandomForest}}, \code{\link{mgcv}}
}

