% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariaAtlas_covariates.R
\name{malariaAtlas_covariates}
\alias{malariaAtlas_covariates}
\title{Covariate module: malariaAtlas_covariates}
\usage{
malariaAtlas_covariates(surface = "Plasmodium falciparum PR2-10",
  extent = c(32, 48, 3, 14), year = rep(NA, length(surface)))
}
\arguments{
\item{surface}{The surface name. See the title column in 
\code{\link{malariaAtlas::listRaster}} for available raster names.}

\item{extent}{Either a length 4 numeric vector giving min longitude, max longitude, min latitude, max latitude; or an object of class Extent}

\item{year}{Numeric vector of same lengthh as \code{surface}. Which year to use for time varying rasters. NA for the latest year or for static rasters.


Unlike \code{\link{malariaAtlas::getRaster}} you cannot get multiple years at once. To get multiple years of the same surface you would have to do
  \code{surface = c('Plasmodium falciparum PR2-10', 'Plasmodium falciparum PR2-10'), year = c(2010, 2011)}.}
}
\description{
Get coovariate rasters from the Malaria Atlas Project
  database.
}
\details{
See the data explorer \link{www.map.ox.ac.uk/explorer} to view 
  available rasters.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2018-04-19
}

\seealso{
\code{\link{malariaAtlas::getraster}} \code{\link{malariaAtlas::listRaster}}

Other covariate: \code{\link{AirNCEP}},
  \code{\link{Bioclim_future}}, \code{\link{Bioclim}},
  \code{\link{CarolinaWrenRasters}},
  \code{\link{LocalRaster}}, \code{\link{NCEP}},
  \code{\link{NaiveRandomRaster}}, \code{\link{UKAir}},
  \code{\link{UKBioclim}}
}
\author{
Tim Lucas, \email{timcdlucas@gmail.com}
}
\concept{covariate}
