% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariaAtlas_PR.R
\name{malariaAtlas_PR}
\alias{malariaAtlas_PR}
\title{Occurrence module: malariaAtlas_PR}
\usage{
malariaAtlas_PR(country = NULL, ISO = "IDN", extent = NULL,
  species = "Pf", standardise = c(2, 10), PA = TRUE, year = NULL,
  fold = 1)
}
\arguments{
\item{country}{Character vector containing names of desired countries, 
e.g. c("Country1", "Country2", ...) OR = "ALL" (use exactly one of 
country, ISO and extent).}

\item{ISO}{Character vector containing ISO3 code for desired country, 
e.g. c("XXX", "YYY", ...) OR = "ALL" (use exactly one of 
 country, ISO and extent).}

\item{extent}{Numeric vector containing bounding box values for desired
area. In the order xmin, xman, ymin, ymax. (use exactly one of 
 country, ISO and extent).}

\item{species}{'Pf' or 'Pv' for P. falciparum or P. vivax respectively.}

\item{standardise}{Length two integer vector giving the age range to
standardise all the points to. The original surveys were conducted
on different age ranges but this affects PR, so typically they should
be standardised. If PA is TRUE, the sample size will be kept the 
same and the number of positives will be rounded. If NULL, no 
standardisation will be performed.}

\item{PA}{Should the data be transformed from parasite rate and examined
data to presence absence data? 
Many modules only work with presence absence data so this TRUE aids 
compatibility. However, the amount of data will often be very large
with FALSE. A single country may have 50,000 presence absence data 
points.}

\item{year}{Vector of years for which to keep data. NULL keeps all years.}

\item{fold}{Control the validation. If 0, all data is validation data.
If 1, all data is training data.}
}
\description{
Occurrence module to collect malaria parasite rate data
  from the Malaria Atlas Project database.
}
\section{Version}{
 1.0
}

\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\references{
Smith, D. L. et al. Standardizing estimates of the
  Plasmodium falciparum parasite rate. Malaria Journal 6, 131 (2007).

  Gething, Peter W., et al. "A long neglected world malaria map:
  Plasmodium vivax endemicity in 2010." PLoS neglected tropical
  diseases 6.9 (2012): e1814.
}
\seealso{
\code{\link{malariaAtlas::getPR}}

Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceDataFrame}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{Lorem_ipsum_UK}}, \code{\link{NATrees}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{SugarMaple}},
  \code{\link{UKAnophelesPlumbeus}}
}
\author{
Tim Lucas, \email{timcdlucas@gmail.com}
}
