% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariaAtlas_VecOcc.R
\name{malariaAtlas_VecOcc}
\alias{malariaAtlas_VecOcc}
\title{Occurrence module: malariaAtlas_VecOcc}
\usage{
malariaAtlas_VecOcc(country = NULL, ISO = "MMR", extent = NULL,
  species = "Anopheles dirus", year = NULL, fold = 1)
}
\arguments{
\item{country}{Character vector containing names of desired countries, 
e.g. c("Country1", "Country2", ...) OR = "ALL" (use exactly one of 
country, ISO and extent).}

\item{ISO}{Character vector containing ISO3 code for desired country, 
e.g. c("XXX", "YYY", ...) OR = "ALL" (use exactly one of 
 country, ISO and extent).}

\item{extent}{Numeric vector containing bounding box values for desired
area. In the order xmin, xman, ymin, ymax. (use exactly one of 
 country, ISO and extent).}

\item{species}{String containing name of desired species}

\item{year}{Vector of years for which to keep data. NULL keeps all years.}

\item{fold}{Control the validation. If 0, all data is validation data.
If 1, all data is training data.}
}
\description{
Occurrence module to collect malaria vector occurrence data
 from the Malaria Atlas Project database.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2018-09-26
}

\section{Data type}{
 presence-only
}

\seealso{
\code{\link{malariaAtlas::getVecOcc}}

Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceDataFrame}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{Lorem_ipsum_UK}}, \code{\link{NATrees}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{SugarMaple}},
  \code{\link{UKAnophelesPlumbeus}},
  \code{\link{malariaAtlas_PR}}
}
\author{
Suzanne H Keddie, \email{suzanne.keddie@bdi.ox.ac.uk}
}
\concept{occurrence}
