% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MaxEnt.R
\name{MaxEnt}
\alias{MaxEnt}
\title{Model module: MaxEnt}
\usage{
MaxEnt(.df)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}.
\code{.df} is data frame that combines the occurrence data and covariate
data. \code{.df} is passed automatically in workflow from the process
module(s) to the model module(s) and should not be passed by the user.}
}
\description{
Model module to fit MaxEnt model via he dismo package
}
\details{
\strong{In order to fit a MaxEnt model, you must first download the
MaxEnt executable file \code{maxent.jar} from:
\url{http://www.cs.princeton.edu/~schapire/maxent/} and save it under the
java directory in the dismo package in your R package library. You can find
the correct location with this command:
\code{paste0(system.file(package = 'dismo'), '/java/')}}.
Running MaxEnt also requires an up-to-date version of java and the java
development kit (which you may already have installed).
}
\seealso{
\code{\link{dismo::maxent}}

Other model: \code{\link{BiomodModel}}; \code{\link{GBM}};
  \code{\link{LogisticRegression}}; \code{\link{OptGRaF}};
  \code{\link{QuickGRaF}}; \code{\link{RandomForest}};
  \code{\link{mgcv}}
}

