% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxLike.R
\name{MaxLike}
\alias{MaxLike}
\title{Model module: MaxLike}
\usage{
MaxLike(.df)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}. 
\code{.df} is data frame that combines the occurrence data and covariate 
data. \code{.df} is passed automatically in workflow from the process 
module(s) to the model module(s) and should not be passed by the user.}
}
\description{
Model module to fit a simple MaxLike model
}
\details{
The maxlike method aims to infer the parameters of a 
  presence-absence logistic regression model using only presence-background 
  data, in a manner somewhat similar to MaxEnt. The practical utility of this
  model has been questioned (see references), since it makes very strong 
  assumptions about the underlying model which are unlikely to be met in 
  practice. A more feature-rich version of the model is implemented in the 
  \code{\link{maxlike::maxlike}} package, but that implementation only 
  accepts raster/point data as input. This module implements a simple version
  of the model using the logistic link, in a format that interacts with the 
  rest of the zoon system.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence/background
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\references{
J.A. Royle, R.B. Chandler, C. Yackulic, and J.D. Nichols (2012). 
  Likelihood analysis of species occurrence probability from presence-only 
  data for modelling species distributions. Methods in Ecology and Evolution,
  3, 545-554.
  
  Hastie, T. & Fithian, W. (2013) Inference from presence-only data; the 
  ongoing controversy. Ecography, 36, 864-867.
  
  Phillips, S.J. & Elith, J. (2013) On estimating probability of presence 
  from use-availability or presence-background data. Ecology, 94, 1409-1419.
  
  Merow, C. & Silander, J.A. (2014) A comparison of Maxlike and Maxent for
  modelling species distributions. Methods in Ecology and Evolution, 5,
  215-225.
}
\seealso{
\code{\link{maxlike::maxlike}}

Other model: \code{\link{BiomodModel}}, \code{\link{GBM}},
  \code{\link{LogisticRegression}},
  \code{\link{MachineLearn}}, \code{\link{MaxEnt}},
  \code{\link{MyMaxLike}}, \code{\link{OptGRaF}},
  \code{\link{QuickGRaF}}, \code{\link{RandomForest}},
  \code{\link{StochasticLogisticRegression}},
  \code{\link{mgcv}}
}

