% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxNet.R
\name{MaxNet}
\alias{MaxNet}
\title{Model module: MaxNet}
\usage{
MaxNet(.df, features = "default", regmult = 1, clamp_predictions = TRUE,
  prediction_type = "logistic")
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}. 
\code{.df} is data frame that combines the occurrence data and covariate 
data. \code{.df} is passed automatically in workflow from the process 
module(s) to the model module(s) and should not be passed by the user.}

\item{features}{A length-one character vector (i.e. a string) defining the 
types of features to use for all covariates (see details).}

\item{regmult}{A positive scalar giving the multiplier for the degree of 
regularisation. Higher values mean more regularisation.}

\item{clamp_predictions}{Whether to clamp predictions when extrapolating.}

\item{prediction_type}{The scale on which to make predictions (see details for
options).}
}
\description{
Model module to fit MaxEnt models using the maxnet package
}
\details{
The maxnet R package fits MaxEnt models using the glmnet package, 
 which enables efficient fitting of glms with regularization. Unlike MaxEnt, 
 MaxNet does not require the user to download and install the MaxEnt java 
 executable.
 
 \code{features} should be a string including an 'l' for linear features, 'q'
 for quadratic features, 'h' for hinge features, 'p' for pairwise 
 interactions and 't' for threshold features. E.g. to use only linear 
 features and their interactions, either \code{features = 'lp'} or 
 \code{features = 'pl'} would work. the default value of \code{'default'} 
 uses maxnet's default settings, adjusting the set up based on the number of 
 occurrence records \emph{np}: 'l' if np < 10; 'lq' if np < 15; 'lqh' if np <
 80; or 'lqph' if np >= 80. I.e. the default never uses threshold features.
 
 \code{prediction_type} corresponds to the \code{type} argument in the 
 \code{maxnet} predict function. \code{maxnet} enables types: 'link', 
 'exponential', 'cloglog' and 'logistic'. However most output modules expect 
 predictions to be made on the probability scale, for which only 'cloglog'
 and 'logistic' are guaranteed to work.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2016-12-21
}

\section{Data type}{
 presence/background
}

\seealso{
\code{\link{maxnet::maxnet}}

Other model: \code{\link{BiomodModel}},
  \code{\link{Domain}}, \code{\link{GBM}},
  \code{\link{LogisticRegression}},
  \code{\link{MachineLearn}}, \code{\link{MaxEnt}},
  \code{\link{MaxLike}}, \code{\link{MyMaxLike}},
  \code{\link{NullModel}}, \code{\link{OptGRaF}},
  \code{\link{QuickGRaF}}, \code{\link{RandomForest}},
  \code{\link{StochasticLogisticRegression}},
  \code{\link{mgcv}}
}
\author{
Nick Golding, \email{nick.golding.research@gmail.com}
}
