% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MESSMask.R
\name{MESSMask}
\alias{MESSMask}
\title{Process module: MESSMask}
\usage{
MESSMask(.data)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow
function}. \code{.data} is a list of a data frame and a raster object
returned from occurrence modules and covariate modules respectively.
\code{.data} is passed automatically in workflow from the occurrence and
covariate modules to the process module(s) and should not be passed by the
user.}
}
\description{
Calculate the multivariate environmental similarity surface 
  and mask areas with values outside the range of variability in the 
  training region.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2016-03-23
}

\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
\code{\link{dismo::mess}}

Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}

