% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv.R
\name{mgcv}
\alias{mgcv}
\title{Model module: mgcv}
\usage{
mgcv(.df, k = -1, bs = "tp")
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}.
\code{.df} is data frame that combines the occurrence}

\item{k}{the dimension of the basis used to represent the smooth term.
The default depends on the number of variables that the smooth is a
 function of. k should not be less than the dimension of the null space
  of the penalty for the term (see \code{\link{mgcv::null.space.dimension}}),
  but will be reset if it is. See \code{\link{mgcv::choose.k}} for further
  information}

\item{bs}{a two letter character string indicating the (penalized)
smoothing basis to use. (eg "tp" for thin plate regression spline,
 "cr" for cubic regression spline). See \code{\link{mgcv::smooth.terms}}
 for an overview of what is available.}
}
\description{
Model module to fit a generalized additive model using generalized
crossvalidation via the mgcv R package.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence/absence
}

\seealso{
Other model: \code{\link{BiomodModel}},
  \code{\link{Domain}}, \code{\link{GBM}},
  \code{\link{LogisticRegression}},
  \code{\link{MachineLearn}}, \code{\link{MaxEnt}},
  \code{\link{MaxLike}}, \code{\link{MaxNet}},
  \code{\link{MyMaxLike}}, \code{\link{OptGRaF}},
  \code{\link{QuickGRaF}}, \code{\link{RandomForest}},
  \code{\link{StochasticLogisticRegression}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
