% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modfun.R
\name{modfun}
\alias{modfun}
\title{Class modfun}
\usage{
modfun(x)
}
\arguments{
\item{x}{(function)}
}
\description{
The constructor and a print method for the class 'modfun', a function in a
module. The only purpose of this class is to have a print method for ad-hoc
documentation.
}
\examples{
fun <- function() {
  ## This is documentation
  ## Triggered by '##'
  NULL
}

modfun(fun)

}

