% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-coercion.R
\name{as.module}
\alias{as.module}
\title{Coercion for Modules}
\usage{
as.module(x, topEncl = baseenv(), ...)
}
\arguments{
\item{x}{something which can be coerced into a module. \code{character} are
interpreted as file names.}

\item{topEncl}{(environment) the root of the local search path.}

\item{...}{arguments passed to \link{parse}}
}
\description{
Interfaces to and from modules.
}

