% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveRandomPresence.R
\name{NaiveRandomPresence}
\alias{NaiveRandomPresence}
\title{Naive Random Presence}
\usage{
NaiveRandomPresence(extent = c(-10, 10, 45, 65), n = 1000, seed = NULL,
  projection = NULL)
}
\arguments{
\item{extent}{A numeric vector of length 4 giving the coordinates of the rectangular region within which to create the random points. order: xmin, xmax, ymin, ymax. By default the extent of the UK.}

\item{n}{Numeric - the number of points to create}

\item{seed}{Used with set.seed to set a seed. Default NULL, no seed is used}

\item{projection}{The projection for your points (as a proj4string)}
}
\description{
Creates a random occurrence dataset within the extent given
}
\details{
Points are created entirely at random using runif
}
\section{Data type}{
 presence-only
}

\section{Version}{
 0
}

\section{Date submitted}{
  2016-05-31
}

\seealso{
Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{Lorem_ipsum_UK}}, \code{\link{NATrees}},
  \code{\link{NBNdataByName}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{SpOcc}}, \code{\link{SugarMaple}},
  \code{\link{UKAnophelesPlumbeus}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
