% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveRandomRaster.R
\name{NaiveRandomRaster}
\alias{NaiveRandomRaster}
\title{Naive Random Raster}
\usage{
NaiveRandomRaster(extent = c(-10, 10, 45, 65), res = 0.5, seed = NULL,
  projection = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{extent}{A numeric vector of length 4 giving the coordinates of the rectangular region within which to create the raaster. order: xmin, xmax, ymin, ymax. By default the extent of the UK}

\item{res}{Numeric giving the resolution of the raster to produce. 1 is one cell per degree, 0.5 means one cell per half degree.}

\item{seed}{Used with set.seed to set a seed. Default NULL, no seed is used}

\item{projection}{The projecdtion for your raster (as a proj4string), defaults to lat/long}
}
\description{
Creates a random raster layer within the extent given
}
\details{

}
\section{Version}{
 0
}

\section{Date submitted}{
  2016-06-02
}

\seealso{
Other covariate: \code{\link{AirNCEP}},
  \code{\link{Bioclim_future}}, \code{\link{Bioclim}},
  \code{\link{CarolinaWrenRasters}},
  \code{\link{LocalRaster}}, \code{\link{NCEP}},
  \code{\link{UKAir}}, \code{\link{UKBioclim}},
  \code{\link{malariaAtlas_covariate}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
