% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NATrees.R
\name{NATrees}
\alias{NATrees}
\title{Occurrences for 21 tree species from North America}
\usage{
NATrees(dateRange = NULL, species = "acer_saccharum")
}
\arguments{
\item{dateRange}{Optional 2-item numeric vector giving the years to use; default is the full range of the data (1960-2013)}

\item{species}{Which species should be extracted, with format genus_species (all lowercase); allowable values are:\nabies_balsamea\nacer_rubrum\nacer_saccharum\nbetula_alleghaniensis\nbetula_papyrifera\nfagus_grandifolia\nfraxinus_americana\nfraxinus_nigra\nlarix_laricina\npicea_glauca\npicea_mariana\npicea_rubens\npinus_banksiana\npinus_resinosa\npinus_strobus\npopulus_grandidentata\npopulous_tremuloides\nquercus_macrocarpa\nquercus_rubra\nthuja_occidentalis\ntsuga_canadensis}
}
\description{
Retrieves occurrences for 21 trees from eastern North America. Data originally sourced from USGS FIA forest inventory database, collected in the QUICC-FOR project ($URL). This dataset was originally published in $REF. Extent for this dataset is c(-97,-67,25,50)
}
\details{

}
\section{Data type}{
 presence/absence
}

\section{Version}{
 0.5
}

\section{Date submitted}{
  2016-06-16
}
\author{
Matt Talluto, \email{mtalluto@gmail.com}
}
\seealso{
Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{Lorem_ipsum_UK}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{SugarMaple}},
  \code{\link{UKAnophelesPlumbeus}}
}

