% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBNdataByName.R
\name{NBNdataByName}
\alias{NBNdataByName}
\title{NBN Data by Common Name}
\usage{
NBNdataByName(species = "Myotis daubentonii", username = "ZOON",
  password = "zoonproject", datasets = NULL, startYear = 2000,
  endYear = 2000, VC = NULL, group = NULL, gridRef = NULL,
  acceptTandC = FALSE, silent = FALSE, attributes = FALSE)
}
\arguments{
\item{species}{The latin name of a species to search for, or a vector of such names.}

\item{username}{your NBN username}

\item{password}{your NBN password}

\item{datasets}{a list of dataset keys which are strings of 8 alphanumeric characters. Look up datasets here: https://data.nbn.org.uk/Datasets}

\item{startYear}{a 4 digit integer year}

\item{endYear}{a 4 digit integer year}

\item{VC}{a string giving a vice-county name (see listVCs in the rnbn package)}

\item{group}{a string giving the name of a group (see listGroups in the rnbn package). Using group will retireve data for all TVKs in this group. for example using group "reptile" will search using over 150 TVKs including TVKs for higher taxonomic groups such families within reptiles. Therefore it may be preferrable to search using a vector of species}

\item{gridRef}{a string giving a UK gridreference in which to search for occurrences}

\item{acceptTandC}{if set to TRUE you accept the NBN gateway terms and conditions and privacy policy. These can be found at https://data.nbn.org.uk/Terms. Accepting the terms and conditions supresses the corresponding warning message.}

\item{silent}{if TRUE information returned to console is reduced}

\item{attributes}{if FALSE then attribute data is not returned, this may improve the speed of large requests.}
}
\description{
This module supplies occurrence data from the UK NBN gateway. To get access to data you must first register at https://data.nbn.org.uk/User/Register. You will need your username and password when running this module. You can specify the data to retrieve by dataset, species, time, location and/or group.
}
\details{
For more detials see the rnbn package
}
\section{Data type}{
 presence-only
}

\section{Version}{
 0.2
}

\section{Date submitted}{
  2016-06-16
}

\seealso{
Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{Lorem_ipsum_UK}}, \code{\link{NATrees}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{SugarMaple}},
  \code{\link{UKAnophelesPlumbeus}}
}
\author{
Tom August, \email{tomaug@ceh.ac.uk}
}
