% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NCEP}
\alias{NCEP}
\title{covariate module to grab coarse resolution environmental data from NCEP}
\arguments{
\item{extent}{A numeric vector of length 4 giving the coordinates of the
rectangular region within which to carry out the analysis, in the
order: xmin, xmax, ymin, ymax.}

\item{variables}{Character vector of which variables to select.
See NCEP.gather for more options but basic options are
      *‘air’*    Air Temperature                 deg K
      *‘hgt’*    Geopotential Height             m
      *‘rhum’*   Relative Humidity               %
      *‘shum’*   Specific Humidity               kg/kg
      *‘omega’*  Omega [Vertical Velocity]       Pascal/s
      *‘uwnd’*   U-Wind Component [East/West]    m/s
      *‘vwnd’*   V-Wind Component [North/South]  m/s}
}
\value{
a Raster* object (class from the raster package) with the gridded
 covariates used to train and predict from the SDM.
}
\description{
covariate module to grab coarse resolution environmental data from NCEP
}
\seealso{
\code{\link{RNCEP::NCEP.gather}}
}

