% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NoProcess.R
\name{NoProcess}
\alias{NoProcess}
\title{Process module: NoProcess}
\usage{
NoProcess(data)
}
\value{
a Raster* object (class from the raster package) with the gridded
     covariates used to train and predict from the SDM.
}
\description{
Process module that does nothing. A place holder for if nothing should be
done to the data before modelling.
}

