% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OneHundredBackground.R
\name{OneHundredBackground}
\alias{OneHundredBackground}
\title{Process module: OneHundredBackground}
\usage{
OneHundredBackground(data)
}
\description{
Process module to generate up to 100 background records at random in
     cells of ras and return these along with the presence only data.
}

