% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OptGRaF.R
\name{OptGRaF}
\alias{OptGRaF}
\title{Model module: OptGRaF}
\usage{
OptGRaF(.df)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}. \code{.df} is data frame that combines the occurrence data and covariate data. \code{.df} is passed automatically in workflow from the process module(s) to the model module(s) and should not be passed by the user.}
}
\description{
Model module to fit a (slow) GRaF model, with parameter optimisation.
}
\seealso{
Other model: \code{\link{BiomodModel}}; \code{\link{GBM}};
  \code{\link{LogisticRegression}}; \code{\link{MaxEnt}};
  \code{\link{QuickGRaF}}; \code{\link{RandomForest}};
  \code{\link{mgcv}}
}

