% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PartitionDisc.R
\name{PartitionDisc}
\alias{PartitionDisc}
\title{PartitionDisc: Leave-one-disc-out cross-validation}
\usage{
PartitionDisc(data, radius = 2, buffer = 1)
}
\arguments{
\item{radius}{Numeric the radius of the disc, in degrees}

\item{buffer}{Numeric, the size of the buffer zone in degrees}
}
\description{
This process module partitions the sample into training and tests set by selecting circular test areas (possibly surrounded by an exclusion buffer) and using the remaining samples as training samples. See function partition.disc in package sperrorest for more details
}
\details{
Module type: Process
}
\author{
Tom August'

tomaug@ceh.ac.uk'
}

