% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictNewRasterMap.R
\name{PredictNewRasterMap}
\alias{PredictNewRasterMap}
\title{Output module: PredictNewRasterMap}
\usage{
PredictNewRasterMap(.model, .ras, newraster = .ras,
  extent = as.vector(.ras@extent), plot = TRUE, ...)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{newraster}{Raster* object to make predictions onto.}

\item{extent}{Numeric vector of length 4 giving (in this order) the min x, max x, min y, max y, the the new area to be mapped. Must be within the covariate data collected in the covariate module.}

\item{plot}{Logical, if \code{TRUE} (default) the prediction map is plotted.}

\item{...}{further arguments to be passed to \code{\link[raster]{plot}}.}
}
\description{
Output module. Predict a model across a new area or time period.
}
\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\section{Version}{
 1.0.1
}

\section{Date submitted}{
  2016-06-15
}

\seealso{
Other output: \code{\link{Appify}},
  \code{\link{CoefficientPlot}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{InteractiveOccurrenceMap}},
  \code{\link{NoOutput}}, \code{\link{PairPlot}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}}, \code{\link{ResponsePlot}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
Zoon developers (modified by F. Rodriguez-Sanchez), \email{zoonproject@gmail.com}
}
