% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintMap.R
\name{PrintMap}
\alias{PrintMap}
\title{Output module: PrintMap}
\usage{
PrintMap(.model, .ras, plot = TRUE, dir = NULL)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{plot}{If \code{TRUE} the plot will be displayed in the device}

\item{dir}{Directory where plots are saved. If \code{NULL} (default) then plots are
not saved.}
}
\value{
A Raster object giving the probabilistic model predictions for each
cell of covariate raster layer
}
\description{
Plot a map of predicted surface.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other output: \code{\link{Figshare}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewAreaMap}},
  \code{\link{ResponseCurve}}
}

