% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintOccurrenceMap.R
\name{PrintOccurrenceMap}
\alias{PrintOccurrenceMap}
\title{Output module: PrintOccurrenceMap}
\usage{
PrintOccurrenceMap(.model, .ras, plot = TRUE, dir = NULL, filename = NULL,
  size = c(480, 480), res = 72, ...)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{plot}{If \code{TRUE} the plot will be displayed in the device}

\item{dir}{Directory where plots are saved. If both \code{dir} and \code{filename} are \code{NULL} (default) then plots are
not saved.}

\item{filename}{The name to be given to the output as a character, don't include a file extension. If both \code{dir} and \code{filename} are \code{NULL} (default) then plots are
not saved.}

\item{size}{A vector containing the width and height of the output figure when writing to a png file. Example: c(800,600).}

\item{res}{The output resolution in ppi when writing to a png file.}

\item{...}{Parameters passed to sp::spplot, useful for setting title and axis labels e.g. \code{xlab = 'Axis Label', main = 'My Plot Title'}}
}
\value{
A Raster object giving the study area
}
\description{
Plot a map of only occurrence data over the study area.
}
\details{
Based on the PrintMap module
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2017-03-07
}

\seealso{
Other output: \code{\link{Appify}},
<<<<<<< HEAD
  \code{\link{CoefficientPlot}},
=======
  \code{\link{GenerateCovariateReport}},
>>>>>>> upstream/master
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{InteractiveOccurrenceMap}},
  \code{\link{NoOutput}}, \code{\link{PairPlot}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{ROCcurve}},
  \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}}, \code{\link{ResponsePlot}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
Nick Golding, \email{nick.golding.research@gmail.com}
}
