% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{RandomForest}
\alias{RandomForest}
\title{Model module: RandomForest}
\usage{
RandomForest(.df)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}. \code{.df} is data frame that combines the occurrence data and covariate data. \code{.df} is passed automatically in workflow from the process module(s) to the model module(s) and should not be passed by the user.}
}
\value{
A model object with a valid predict method
}
\description{
Model module to fit a simple RandomForest classification model
}
\seealso{
Other model: \code{\link{BiomodModel}}; \code{\link{GBM}};
  \code{\link{LogisticRegression}};
  \code{\link{MachineLearn}}; \code{\link{MaxEnt}};
  \code{\link{OptGRaF}}; \code{\link{QuickGRaF}};
  \code{\link{mgcv}}
}

