% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityPlot.R
\name{ReliabilityPlot}
\alias{ReliabilityPlot}
\title{Produces are relaibility plot with associate linear regression analysis}
\usage{
ReliabilityPlot(.model, .ras, n_bins = 20)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{n_bins}{An integer greater than 1 indicating the number of bins for predictions and obsertions}
}
\description{
Returns a reliability plot of the discritised predictions versus obserations as well as the a vector of the intercept, gradient, and r-squared for further interpretation
}
\details{
We can analyse the linear model of predictions versus observations for bias in our predictions. A gradient on 1 and and intercept of 0 indicate a perfect fit - the predictions equal the observations. An intercept less than 0 indicated that, on average, our model underpredicts. Greater than 0 indicates over-prediction, on average.
}
\section{Data type}{
 presence-only, presence/absence, presence/background
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2018-02-27
}

\seealso{
Other output: \code{\link{AIC}}, \code{\link{AUC}},
  \code{\link{Appify}}, \code{\link{CoefficientPlot}},
  \code{\link{CovHistograms}}, \code{\link{DataSummary}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{InteractiveOccurrenceMap}},
  \code{\link{NoOutput}}, \code{\link{PairPlot}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}}, \code{\link{ResponsePlot}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
Liz Martin, \email{lizmartinresearch@gmail.com}
}
