% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveNAs.R
\name{RemoveNAs}
\alias{RemoveNAs}
\title{Process module: RemoveNAs}
\usage{
RemoveNAs(.data, responseOnly = FALSE, ignore = NULL)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{responseOnly}{Logical indicating whether to only remove rows with NAs in the response column.}

\item{ignore}{Character vector of columns to ignore i.e. columns that can have NAs without being removed.}
}
\description{
Process module that removes data containing NAs. 
Remove either data with NAs in any column or only data with NAs
in the response. As a number of model modules cannot handle NAs, this
may be necessary.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2018-04-11
}

\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}},
  \code{\link{LonLatToCovariates}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}},
  \code{\link{spThin}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
