% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseCurveViz.R
\name{ResponseCurveViz}
\alias{ResponseCurveViz}
\title{Conditional Plots}
\usage{
ResponseCurveViz(.model, .ras, cov = NULL)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{cov}{Covariates}
}
\description{
Basic conditional plots in one column
}
\details{
Just starting out, much more to do
}
\section{Data type}{
 presence-only
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2016-06-15
}

\seealso{
Other output: \code{\link{Appify}},
  \code{\link{CoefficientPlot}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}}, \code{\link{NoOutput}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurve}},
  \code{\link{ResponsePlot}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
G. McInerny, \email{gmcinerny@hotmail.com}
}
