% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponsePlot.R
\name{ResponsePlot}
\alias{ResponsePlot}
\title{Plot predicted response curves along covariate gradients}
\usage{
ResponsePlot(.model, .ras, cov = NULL)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{cov}{covariate name}
}
\description{
This module plots predicted response curves along covariate gradients for all or specified covariates
}
\details{
The module outputs a line graph of probability of occupancy (relative for presence-only and presence background data, absolute for presence-absence data). cov accepts a numeric and character strings. When cov = NULL, the module plots line graphs for each covariate in the zoon workflow iteratively. Currently only implemented for presence-only, presence-background, and presence-absence data.
}
\section{Data type}{
 presence-only, presence/absence
}

\section{Version}{
 0.2
}

\section{Date submitted}{
  2017-05-03
}

\seealso{
Other output: \code{\link{Appify}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}}, \code{\link{NoOutput}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
Liz Martin, \email{emartin@student.unimelb.edu.au}
}
