% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sameTimePlaceMap}
\alias{sameTimePlaceMap}
\title{A function for outputing the raster of the predictions from an analysis}
\usage{
sameTimePlaceMap(model, ras)
}
\arguments{
\item{model}{A model object, the output from a model module}

\item{ras}{A Raster* object, the output from a covariate module}
}
\value{
A Raster object giving the probabilistic model predictions for each
     cell of ras
}
\description{
A function for outputing the raster of the predictions from an analysis
}

