% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spThin.R
\name{spThin}
\alias{spThin}
\title{Spatial thinning of Presence-only Data}
\usage{
spThin(.data, thin = 50)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{thin}{Thinning parameter - the required minimum distance (in kilometres) between points after applying the thinning procedure}
}
\description{
Apply the stochastic spatial thinning algorithm implemented in the spThin package to presence data in a presence-background dataset
}
\details{
Full details of the algorithm are available in the open-access article by Aiello-Lammens et al. (2015): dx.doi.org/10.1111/ecog.01132
}
\section{Data type}{
 presence-only
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2017-11-23
}

\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}},
  \code{\link{LonLatToCovariates}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}}, \code{\link{RemoveNAs}},
  \code{\link{StandardiseCov}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}
\author{
Nick Golding, \email{nick.golding.research@gmail.com}
}
