% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardiseCov.R
\name{StandardiseCov}
\alias{StandardiseCov}
\title{Scale numeric covariates}
\usage{
StandardiseCov(.data, Gelman = FALSE, exclude = NULL)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{Gelman}{Option to divide by two rather than one standard deviation.}

\item{exclude}{Variable names of numeric variables to exclude from standardisation}
}
\description{
The module by default scales all numeric covariates. By default conducts a regular standardisation: minus the mean and divide by the standard deviation. If Gelman = TRUE then divides by 2 sd as suggested by Gelman. Excluded covariates are not standardised.
}
\details{

}
\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\section{Version}{
 0
}

\section{Date submitted}{
  2016-06-16
}
\author{
Alison Johnston & Carsten F. Dormann, \email{alison.johnston@bto.org}
}
\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}

