% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochasticLogisticRegression.R
\name{StochasticLogisticRegression}
\alias{StochasticLogisticRegression}
\title{StochasticLogisticRegression}
\usage{
StochasticLogisticRegression(.df)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}. \code{.df} is data frame that combines the occurrence data and covariate data. \code{.df} is passed automatically in workflow from the process module(s) to the model module(s) and should not be passed by the user.}
}
\description{
Fits a simple logistic regression model, but then samples a single draw of the model coefficients (with probability proportional to their likelihood) and uses these for future prediction
}
\details{
Coefficients are simulated from the likelihood density using the approximate hessian matrix, under an assumption of multivariate normality. This module is intended for use in a Monte Carlo simulation procedure to propagate uncertainty through an analysis. It is not intended to be used on its own!
}
\section{Data type}{
 presence/absence
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2016-06-16
}

\seealso{
Other model: \code{\link{BiomodModel}},
  \code{\link{Domain}}, \code{\link{GBM}},
  \code{\link{LogisticRegression}},
  \code{\link{MachineLearn}}, \code{\link{MaxEnt}},
  \code{\link{MaxLike}}, \code{\link{MaxNet}},
  \code{\link{MyMaxLike}}, \code{\link{NullModel}},
  \code{\link{OptGRaF}}, \code{\link{QuickGRaF}},
  \code{\link{RandomForest}}, \code{\link{mgcv}}
}
\author{
Nick Golding, \email{nick.golding.research@gmail.com}
}
\concept{model}
