% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleOccurrence.R
\name{SubsampleOccurrence}
\alias{SubsampleOccurrence}
\title{Process module: SubsampleOccurrence}
\usage{
SubsampleOccurrence(.data, n = 100, replace = TRUE)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow 
function}. \code{.data} is a list of a data frame and a raster object 
returned from occurrence modules and covariate modules respectively. 
\code{.data} is passed automatically in workflow from the occurrence and 
covariate modules to the process module(s) and should not be passed by the 
user.}

\item{n}{the number of points to sample. Defaults to the same number as the
dataset}

\item{replace}{whether to sample with replacement}
}
\description{
Process module to sample a random subset of the occurrence data.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2016-12-21
}

\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}
\author{
Nick Golding, \email{nick.golding.research@gmail.com}
}
