% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SugarMaple.R
\name{SugarMaple}
\alias{SugarMaple}
\title{Sugar Maple distribution in North America}
\usage{
SugarMaple()
}
\description{
Presence/absence of Sugar Maple (\emph{Acer saccharum}) in North America. Each record corresponds to the centroid of a 0.5 degree cell.
}
\details{
The extent of the occurrence data is \code{c(-170, -20, 10, 80)}. These data were used as part of a wider analysis by Morin & Thuiller (2009), and were subsequently analysed and made available by Talluto \emph{et al.}(2016). Morin, X., & Thuiller, W. (2009). Comparing niche-and process-based models to reduce prediction uncertainty in species range shifts under climate change. Ecology, 90(5), 1301–1313. http://doi.org/10.1890/08-0134.1  Talluto, M. V., Boulangeat, I., Ameztegui, A., Aubin, I., Berteaux, D., Butler, A., … Gravel, D. (2016). Cross-scale integration of knowledge for predicting species ranges: A metamodelling framework. Global Ecology and Biogeography, 25(2), 238–249. http://doi.org/10.1111/geb.12395
}
\section{Data type}{
 presence/absence
}

\section{Version}{
 0.2
}

\section{Date submitted}{
  2016-06-15
}

\seealso{
Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceDataFrame}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{Lorem_ipsum_UK}}, \code{\link{NATrees}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{UKAnophelesPlumbeus}}
}
\author{
Matt Talluto & Nick Golding, \email{mtalluto@gmail.com}
}
\concept{occurrence}
