% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetGroupBackground.R
\name{TargetGroupBackground}
\alias{TargetGroupBackground}
\title{Target Group Background Data}
\usage{
TargetGroupBackground(.data, OccurrenceCall = NaiveRandomPresence,
  naOmit = TRUE)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{OccurrenceCall}{A call to an occurrence module in the same format as would be given to the "Occurrence" argument in workflow. Can be a Chain but cannot be a list (instead list TargetGroupBackground).}

\item{naOmit}{Should NA containing rows be removed from the dataset (occurrence and background points combine). Defaults to TRUE}
}
\description{
This module takes as one of its arguements a call to an occurrence module, these occurrence points are then used as background points. This follows the methods outlined by Phillips et al (Phillips, S. J., Dudík, M., Elith, J., Graham, C. H., Lehmann, A., Leathwick, J. and Ferrier, S. (2009), Sample selection bias and presence-only distribution models: implications for background and pseudo-absence data. Ecological Applications, 19: 181–197. doi:10.1890/07-2153.1) as a method to choose background points with similar bias to the target species presence points.
}
\details{

}
\section{Data type}{
 presence-only
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2016-08-22
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}

