% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{Transform}
\alias{Transform}
\title{Apply Transformations to Covariates}
\usage{
Transform(.data, trans = function(x) {     x }, which_cov = NULL,
  replace = TRUE)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.#'}

\item{trans}{a function defining a transformation to apply to each pixel in the covariate(s) raster. This can be a definition (e.g. \code{function(x) x^2} for a square) or a function defined in base R (e.g. \code{log1p} for the log(x + 1) transformation). The default transformation does nothing.#'}

\item{which_cov}{a character vector naming all of the covariates in the RasterStack to which the new transformation should be applied.#'}

\item{replace}{a logical, determining whether to overwrite the values of the named layers with the transformed values (if \code{replace = TRUE}), or to add extra layers with the transformed covariates. If additional layers are added, they will be renamed accoring to the \code{raster} package's renaming conventions. This normally involves adding a '.2' at the end of the name.}
}
\description{
Apply a tranformation function (e.g. square, log, something else) to one or more covariates, pixelwise. These can either overwrite the named covariates, or be added to the covariate set.
}
\details{
This module is essentially a wrapper around the \code{raster} function \code{\link[raster]{calc}}.
}
\section{Data type}{
 presence-only, presence/absence, presence/background, abundance, proportion
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2016-06-15
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{TargetGroupBackground}},
  \code{\link{addInteraction}}
}

