% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKAir.R
\name{UKAir}
\alias{UKAir}
\title{Covariate module: UKAir}
\usage{
UKAir()
}
\description{
Return some UK air temperature data from NCEP. Data is bundled with package and has the extent \code{c(-10, 10, 45, 65)}.
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other covariate: \code{\link{AirNCEP}},
  \code{\link{Bioclim}}, \code{\link{LocalRaster}},
  \code{\link{NCEP}}, \code{\link{UKBioclim}}
}

