% Generated by roxygen2 (4.0.1): do not edit by hand
\name{UKAnophelesPlumbeus}
\alias{UKAir}
\alias{UKAnophelesPlumbeus}
\title{Return some UK air temperature data from NCEP. Data is bundled with package.}
\usage{
UKAir()
}
\description{
Return some UK air temperature data from NCEP. Data is bundled with package.

Return some Anopheles plumbeus data that is bundled with package.
 Data taken from GBIF and is UK extent.
}

