% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_cached}
\alias{use_cached}
\alias{invalidate_cache}
\title{Use Module File - Cached}
\usage{
use_cached(module_file, ...)

invalidate_cache()
}
\arguments{
\item{module_file}{path to module R script file}

\item{...}{parameters passed to `use`}
}
\description{
A cached version of `use`. Only supports module file R scripts.
Useful if module scripts take too long to parse, especially when there are nested modules.

Returned cached modules are copy on modify, so altering elements of a module will not affect other copies else where.
Module function enclosing environments are also recursively cloned to prevent global effects across cached modules.

Cache invalidation is based on modified time of `module_file`.
However, the modified time of nested modules are not checked.
In such case, use `invalidate_cache` for manual cache invalidation.
}
