% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableImportance.R
\name{VariableImportance}
\alias{VariableImportance}
\title{Variable Importance}
\usage{
VariableImportance(.model, .ras, print = TRUE)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{print}{If true then the data are returned to console as well as outputted}
}
\description{
This module outputs a simple report of the coefficents/importance measures from the model
}
\details{
This module identifies the class of the model object and if it has a known return method for this class it will be applied. If the class is not known it will try the default which is to return the summary of the object.
}
\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2016-06-15
}

\seealso{
Other output: \code{\link{AUC}}, \code{\link{Appify}},
  \code{\link{CoefficientPlot}},
  \code{\link{CovHistograms}}, \code{\link{DataSummary}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{InteractiveOccurrenceMap}},
  \code{\link{NoOutput}}, \code{\link{PairPlot}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PredictionCalibration}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ReliabilityPlot}},
  \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}}, \code{\link{ResponsePlot}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}}
}
\author{
Tom August, \email{tomaug@ceh.ac.uk}
}
