% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AirNCEP.R
\name{AirNCEP}
\alias{AirNCEP}
\title{Covariate module: AirNCEP}
\usage{
AirNCEP()
}
\description{
Covariate module to grab a coarse resolution mean air temperature raster from
 January-February 2001-2002 for the UK.
}
\seealso{
\code{\link{RNCEP::NCEP.gather}}
}

