% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AirNCEP}
\alias{AirNCEP}
\title{covariate module to grab a coarse resolution mean air temperature raster from
      January-February 2001-2002 for the given extent.}
\usage{
AirNCEP()
}
\arguments{
\item{extent}{A numeric vector of length 4 giving the coordinates of the
     rectangular region within which to carry out the analysis, in the
     order: xmin, xmax, ymin, ymax.}
}
\value{
a Raster* object (class from the raster package) with the gridded
     covariates used to train and predict from the SDM.
}
\description{
covariate module to grab a coarse resolution mean air temperature raster from
      January-February 2001-2002 for the given extent.
}

