% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AirNCEP}
\alias{AirNCEP}
\title{Covariate module: AirNCEP}
\description{
Covariate module to grab a coarse resolution mean air temperature raster from
 January-February 2001-2002 for the given extent.
}
\seealso{
\code{\link{RNCEP::NCEP.gather}}
}

