% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirNCEP.R
\name{AirNCEP}
\alias{AirNCEP}
\title{Covariate module: AirNCEP}
\usage{
AirNCEP(quiet = TRUE)
}
\arguments{
\item{quiet}{Logical. If TRUE (default) the progress of downloads is not shown}
}
\description{
Covariate module to grab a coarse resolution mean air temperature raster from
 January-February 2001-2002 for the UK.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\seealso{
\code{\link{RNCEP::NCEP.gather}}

Other covariate: \code{\link{Bioclim_future}},
  \code{\link{Bioclim}}, \code{\link{CarolinaWrenRasters}},
  \code{\link{LocalRaster}}, \code{\link{NCEP}},
  \code{\link{NaiveRandomRaster}}, \code{\link{UKAir}},
  \code{\link{UKBioclim}},
  \code{\link{malariaAtlas_covariates}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\concept{covariate}
