% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AirNCEP}
\alias{AirNCEP}
\title{covariate module to grab a coarse resolution mean air temperature raster from
 January-February 2001-2002 for the given extent.}
\arguments{
\item{modelType}{A character vector to describe models to use. Select from
'GLM','GBM','GAM','CTA','ANN','SRE','FDA','MARS','RF','MAXENT'}
}
\value{
a Raster* object (class from the raster package) with the gridded
 covariates used to train and predict from the SDM.

a Raster* object (class from the raster package) with the gridded
     covariates used to train and predict from the SDM.
}
\description{
covariate module to grab a coarse resolution mean air temperature raster from
 January-February 2001-2002 for the given extent.

Model module wrapper for BIOMOD_Modeling()
}
\seealso{
\code{\link{RNCEP::NCEP.gather}}

\code{\link{biomod2::BIOMOD_ModelingOptions}}
}

