% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/AnophelesPlumbeus.R
\name{AnophelesPlumbeus}
\alias{AnophelesPlumbeus}
\title{Occurrence module: AnophelesPlumbeus}
\usage{
AnophelesPlumbeus(extent = c(-10, 10, 45, 65))
}
\arguments{
\item{extent}{A numeric vector of length 4 giving the coordinates of the
      rectangular region within which to carry out the analysis, in the
      order: xmin, xmax, ymin, ymax. By default the extent of the UK.}
}
\description{
Occurrence module to grab *Anopheles plumbeus* (a mosquito) presence
      (presence-only) data from GBIF, in the area bounded by extent.
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other occurrence: \code{\link{CWBZimbabwe}};
  \code{\link{LocalOccurrenceData}}; \code{\link{SpOcc}};
  \code{\link{UKAnophelesPlumbeus}}
}

