% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AnophelesPlumbeus.R
\name{AnophelesPlumbeus}
\alias{AnophelesPlumbeus}
\title{Occurrence module: AnophelesPlumbeus}
\usage{
AnophelesPlumbeus(extent)
}
\arguments{
\item{extent}{A numeric vetor of length 4 giving the coordinates of the
      rectangular region within which to carry out the analysis, in the
      order: xmin, xmax, ymin, ymax.}
}
\description{
Occurrence module to grab *Anopheles plumbeus* (a mosquito) occurrence (i.e.
      presence-only) data from GBIF, in the area bounded by extent.
      Perhaps this should have temporal interval too for future-proofing?
}

