% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Background.R
\name{Background}
\alias{Background}
\title{Process module: Background}
\usage{
Background(.data, n = 100, bias = NULL, seed = NULL)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow
function}. \code{.data} is a list of a data frame and a raster object
returned from occurrence modules and covariate modules respectively.
\code{.data} is passed automatically in workflow from the occurrence and
covariate modules to the process module(s) and should not be passed by the
user.}

\item{n}{the number of background points to sample}

\item{bias}{optional \code{RasterLayer} with cells giving the relative
probability of a background record being sampled there. Alternatively,
a length one numeric giving a radius (in KM) around presence points to take 
background points from. If \code{bias = NULL} (the default) then no 
biasing is applied, and all non-missing cells are equally likely to be selected.}

\item{seed}{Numeric used with \code{\link[base]{set.seed}}}
}
\description{
Process module to generate random (optionally biased) background
  records in cells of the covariate raster and return these along with the
  presence only data.
}
\section{Version}{
 1.1
}

\section{Date submitted}{
 2016-03-23
}

\section{Data type}{
 presence-only
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}

