% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BackgroundAndCrossvalid}
\alias{BackgroundAndCrossvalid}
\title{Process module: BackgroundAndCrossvalid}
\usage{
BackgroundAndCrossvalid(data, k = 5)
}
\arguments{
\item{k}{The number of folds you wish to have. Will later implement a leaveoneout opt}
}
\description{
Process module to generate up to 100 background records at random in
     cells of ras and split all data in k folds for cross validation.
}

