% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackgroundAndCrossvalid.R
\name{BackgroundAndCrossvalid}
\alias{BackgroundAndCrossvalid}
\title{Process module: BackgroundAndCrossvalid}
\usage{
BackgroundAndCrossvalid(.data, k = 5, seed = NULL)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{k}{The number of folds you wish to have. Will later implement a leaveoneout opt}

\item{seed}{Numeric used with \code{\link[base]{set.seed}}}
}
\description{
Process module to generate up to 100 background records at random in
     cells of ras and split all data in k folds for cross validation.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence-only
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}

