<<<<<<< HEAD
% Generated by roxygen2 (4.0.2): do not edit by hand
=======
% Generated by roxygen2 (4.0.1): do not edit by hand
>>>>>>> 7c7e78e360d831dafdb62d23fefdf4086a241211
\name{BackgroundAndCrossvalid}
\alias{BackgroundAndCrossvalid}
\title{Process module: BackgroundAndCrossvalid}
\usage{
BackgroundAndCrossvalid(occurrence, ras, k = 5)
}
<<<<<<< HEAD
=======
\arguments{
\item{k}{The number of folds you wish to have. Will later implement a 'leaveoneout' opt}
}
>>>>>>> 7c7e78e360d831dafdb62d23fefdf4086a241211
\description{
Process module to generate up to 100 background records at random in
     cells of ras and split all data in k folds for cross validation.
}

