% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Bioclim}
\alias{Bioclim}
\title{Covariate module: Bioclim}
\usage{
Bioclim(extent, resolution = 5)
}
\arguments{
\item{extent}{Length 4 character vector giving min lat, max lat, min long, max long.}

\item{resolution}{Resolution in minutes. Must be one of 2.5, 5 or 10. Default is 5.}
}
\description{
Get worldclim environment data. Downloads then stores locally.
}

