% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioclim.R
\name{Bioclim}
\alias{Bioclim}
\title{Covariate module: Bioclim}
\usage{
Bioclim(extent = c(-180, 180, -90, 90), resolution = 10, layers = 1:19)
}
\arguments{
\item{extent}{Length 4 numeric vector giving min longitude, max longitude, min latitude, max latitude.}

\item{resolution}{Resolution in minutes. Must be one of 2.5, 5 or 10. Default is 10.}

\item{layers}{which bioclim layers to obtain, a vector of integers between 1 and 19.}
}
\description{
Get worldclim environment data. Downloads then stores locally.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other covariate: \code{\link{AirNCEP}},
  \code{\link{CarolinaWrenRasters}},
  \code{\link{LocalRaster}}, \code{\link{NCEP}},
  \code{\link{UKAir}}, \code{\link{UKBioclim}}
}

