% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BioclimNewArea.R
\name{BioclimNewArea}
\alias{BioclimNewArea}
\title{Covariate module: BioclimNewArea}
\usage{
BioclimNewArea(trainExtent, predictExtent, resolution = 5)
}
\arguments{
\item{trainExtent}{Length 4 character vector giving min longitude, max longitude, min latitude, max latitude.}

\item{predictExtent}{Length 4 character vector giving min longitude, max longitude, min latitude, max latitude.}

\item{resolution}{Resolution in minutes. Must be one of 2.5, 5 or 10. Default is 5.}
}
\description{
Get worldclim environment data. Downloads then stores locally.
}

