% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BiomodModel.R
\name{BiomodModel}
\alias{BiomodModel}
\title{Model module: BiomodModel}
\usage{
BiomodModel(.df, modelType)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}.
\code{.df} is data frame that combines the occurrence data and covariate
 data. \code{.df} is passed automatically in workflow from the process
 module(s) to the model module(s) and should not be passed by the user.}

\item{modelType}{A character vector to describe models to use. Select from
'GLM','GBM','GAM','CTA','ANN','SRE','FDA','MARS','RF','MAXENT'}
}
\description{
Model module wrapper for the biomod2 function BIOMOD_Modeling()
}
\seealso{
\code{\link{biomod2::BIOMOD_ModelingOptions}}

Other model: \code{\link{GBM}};
  \code{\link{LogisticRegression}}; \code{\link{MaxEnt}};
  \code{\link{OptGRaF}}; \code{\link{QuickGRaF}};
  \code{\link{RandomForest}}; \code{\link{mgcv}}
}

