% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomodModel.R
\name{BiomodModel}
\alias{BiomodModel}
\title{Model module: BiomodModel}
\usage{
BiomodModel(.df, modelType = "GLM")
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}.
\code{.df} is data frame that combines the occurrence data and covariate
 data. \code{.df} is passed automatically in workflow from the process
 module(s) to the model module(s) and should not be passed by the user.}

\item{modelType}{A character to describe models to use. Select from
'GLM','GBM','GAM','CTA','ANN','SRE','FDA','MARS','RF','MAXENT'}
}
\description{
Model module wrapper for the biomod2 function BIOMOD_Modeling()
}
\details{
In order to fit a MaxEnt model, you must first download the
MaxEnt executable file \code{maxent.jar} and save it in the correct location.
The zoon function \code{GetMaxEnt} can orchestrate this for you.
Running MaxEnt also requires an up-to-date version of java
(which you may already have installed).
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence/background, presence/absence, abundance, proportion
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
\code{\link{biomod2::BIOMOD_ModelingOptions}}

Other model: \code{\link{GBM}},
  \code{\link{LogisticRegression}},
  \code{\link{MachineLearn}}, \code{\link{MaxEnt}},
  \code{\link{MaxLike}}, \code{\link{MaxNet}},
  \code{\link{MyMaxLike}}, \code{\link{OptGRaF}},
  \code{\link{QuickGRaF}}, \code{\link{RandomForest}},
  \code{\link{StochasticLogisticRegression}},
  \code{\link{mgcv}}
}

