% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarolinaWrenPA.R
\name{CarolinaWrenPA}
\alias{CarolinaWrenPA}
\title{Presence/absence data for the Carolina Wren in the USA}
\usage{
CarolinaWrenPA()
}
\description{
Presence/absence of the Carolina Wren \emph{Thryothorus ludovicianus} at Breeding Bird Survey transects in 2006. These data are made available in the maxlike R package. See \code{\link[maxlike]{carw.data}} for more details. This module loads them data transform the coordinates to lat/longs and formats the data for use in a zoon workflow.
}
\details{

}
\section{Data type}{
 presence/absence
}

\section{Version}{
 0
}

\section{Date submitted}{
  2016-05-21
}
\author{
Nick Golding, \email{nick.golding.research@gmail.com}
}
\seealso{
Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{UKAnophelesPlumbeus}}
}

