% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Crossvalidate.R
\name{Crossvalidate}
\alias{Crossvalidate}
\title{Process module: Crossvalidate}
\usage{
Crossvalidate(.data, k = 5)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{k}{Positive integer number of folds to split the data into. Default is 5.}
}
\description{
Run k fold crossvalidation. If presence absence, split presences and absences separately so folds have equally balanced data. Otherwise just sample.
}
\seealso{
Other process: \code{\link{BackgroundAndCrossvalid}};
  \code{\link{Background}}; \code{\link{BiasBackground}};
  \code{\link{NoProcess}};
  \code{\link{OneHundredBackground}};
  \code{\link{OneThousandBackground}};
  \code{\link{PartitionDisc}}
}

