% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Crossvalidate}
\alias{Crossvalidate}
\title{Process module: Crossvalidate}
\usage{
Crossvalidate(data, k = 5)
}
\arguments{
\item{k}{Positive integer number of folds to split the data into. Default is 5.}
}
\description{
Run k fold crossvalidation. If presence absence, split presences and absences separately so folds have equally balanced data. Otherwise just sample.
}

