% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossvalidate.R
\name{Crossvalidate}
\alias{Crossvalidate}
\title{Process module: Crossvalidate}
\usage{
Crossvalidate(.data, k = 5, seed = NULL)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{k}{Positive integer number of folds to split the data into. Default is 5.}

\item{seed}{Numeric used with \code{\link[base]{set.seed}}}
}
\description{
Run k fold crossvalidation. If presence absence, split presences and absences separately so folds have equally balanced data. Otherwise just sample.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{JitterOccurrence}},
  \code{\link{MESSMask}}, \code{\link{NoProcess}},
  \code{\link{OneHundredBackground}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}},
  \code{\link{StandardiseCov}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
