% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCovariateReport.R
\name{GenerateCovariateReport}
\alias{GenerateCovariateReport}
\title{Generate a data profiling report}
\usage{
GenerateCovariateReport(.model, .ras)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}
}
\description{
This output module generates a report outlining the results of data profiling. This includes data structure, missing values, and correlation testing.
}
\details{
The results are presented as an html output generated using Rmarkdown that opens in your internet browser
}
\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\section{Version}{
 0.1
}

\section{Date submitted}{
  2017-04-24
}

\seealso{
Other output: \code{\link{Appify}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}}, \code{\link{NoOutput}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
David Wilkinson, \email{davidpw@student.unimelb.edu.au}
}
