% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveMap.R
\name{InteractiveMap}
\alias{InteractiveMap}
\title{Output module: InteractiveMap}
\usage{
InteractiveMap(.model, .ras)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.#'}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}
}
\description{
Plot a zoomable and scrollable map of the predicted distribution
and training data. Clicking on a point reveals additional information.
}
\section{Version}{
 1.1
}

\section{Date submitted}{
 2017-04-05
}

\seealso{
Other output: \code{\link{Appify}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveOccurrenceMap}},
  \code{\link{NoOutput}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
