% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LocalData}
\alias{LocalData}
\title{Occurrence module: LocalData}
\usage{
LocalData(filename, occurrenceType)
}
\arguments{
\item{filename}{The path to the file. Currently assumed a .csv file.}

\item{occurrenceType}{What type data is it?
 One of 'presence', 'presence/absence', 'abundance')}
}
\description{
Occurrence module to format local occurrence data to be used with zoon.
}
\seealso{
\code{\link{read.csv}}
}

