% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LocalOccurrenceData}
\alias{LocalOccurrenceData}
\title{Occurrence module: LocalOccurrenceData}
\usage{
LocalOccurrenceData(filename, occurrenceType, externalValidation = FALSE)
}
\arguments{
\item{filename}{The path to the file. Currently assumes a .csv file.}

\item{occurrenceType}{What type data is it? One of 'presence', 'presence/absence', 'abundance'}

\item{externalValidation}{Logical indicating whether this data is external validation data. Cross validation is handled elsewhere and this argument should be left as FALSE for data that will be used in cross validation.}
}
\description{
Occurrence module to format local occurrence data to be used with zoon. Must be a .csv file with three columns longitude, latitude and value in that order.
}

