% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LocalRaster}
\alias{LocalRaster}
\title{Covariate module: LocalRaster}
\usage{
LocalRaster(filenames)
}
\arguments{
\item{filenames}{Either a string of the filename of the raster layer, or a list of strings of filenames to rasters that will be stacked.}
}
\description{
Read in a single raster or a list of rasters and stack them.
}

