% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LocalRaster.R
\name{LocalRaster}
\alias{LocalRaster}
\title{Covariate module: LocalRaster}
\usage{
LocalRaster(filenames)
}
\arguments{
\item{filenames}{Either a string of the filename of the raster layer, or a list of strings of filenames to rasters that will be stacked.}
}
\description{
LocalRaster: Read in a single raster or a list of rasters and stack them.
}
\details{
Read in a single raster or a list of rasters and stack them.
}
\author{
Tim Lucas

Email: timcdlucas@gmail.com
}
\seealso{
Other covariate: \code{\link{AirNCEP}};
  \code{\link{Bioclim}}; \code{\link{NCEP}};
  \code{\link{UKAir}}; \code{\link{UKBioclim}}
}

