% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalRaster.R
\name{LocalRaster}
\alias{LocalRaster}
\title{Covariate module: LocalRaster}
\usage{
LocalRaster(rasters = "myRaster")
}
\arguments{
\item{rasters}{Either a string of the filename of the raster layer, a list or vector of strings of filenames to rasters that will be stacked, a RasterLayer or a RasterStack object.}
}
\description{
LocalRaster: Read in a single raster or a list of rasters and stack them.
}
\details{
Read in a single raster or a list of rasters and stack them.
}
\section{Version}{
 1.1
}

\section{Date submitted}{
 2016-03-04
}

\seealso{
Other covariate: \code{\link{AirNCEP}},
  \code{\link{Bioclim_future}}, \code{\link{Bioclim}},
  \code{\link{CarolinaWrenRasters}}, \code{\link{NCEP}},
  \code{\link{NaiveRandomRaster}}, \code{\link{UKAir}},
  \code{\link{UKBioclim}},
  \code{\link{malariaAtlas_covariates}}
}
\author{
Tim Lucas, Samuel Bosch, \email{timcdlucas@gmail.com}
}
