% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LocalRaster2}
\alias{LocalRaster2}
\title{Covariate module: LocalRaster2}
\usage{
LocalRaster2(filenames)
}
\arguments{
\item{filenames}{Either a string of the filename of the raster layer, or a list of strings of filenames to rasters that will be stacked.}
}
\description{
Read in a single raster or a list of rasters and stack them.
}

