% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{LogisticRegression}
\alias{LogisticRegression}
\title{Model module: LogisticRegression}
\usage{
LogisticRegression(.df)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}. \code{.df} is data frame that combines the occurrence data and covariate data. \code{.df} is passed automatically in workflow from the process module(s) to the model module(s) and should not be passed by the user.}
}
\description{
Model module to fit a simple logistic regression model
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
\code{\link{glm}}

Other model: \code{\link{BiomodModel}}; \code{\link{GBM}};
  \code{\link{MachineLearn}}; \code{\link{MaxEnt}};
  \code{\link{OptGRaF}}; \code{\link{QuickGRaF}};
  \code{\link{RandomForest}}; \code{\link{mgcv}}
}

