% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/MaxEnt.R
\name{MaxEnt}
\alias{MaxEnt}
\title{Model module: MaxEnt}
\usage{
MaxEnt(.df, args = "")
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}.
\code{.df} is data frame that combines the occurrence data and covariate
data. \code{.df} is passed automatically in workflow from the process
module(s) to the model module(s) and should not be passed by the user.}

\item{args}{Additional arguments to pass to MaxEnt to modify model fitting,
in the form of flags to the java executable.
Unfortunately these are a bit cryptic and the maxent
help files, which list all of the options, aren't available on the web.
You could try the MaxEnt graphical user interface, otherwise there is a list
(which may or may not be up-to-date) in this forum:
\url{https://groups.google.com/d/msg/maxent/yRBlvZ1_9rQ/Fj8Two0lmHIJ}.}
}
\description{
Model module to fit MaxEnt model via the dismo package
}
\details{
In order to fit a MaxEnt model, you must first download the
MaxEnt executable file \code{maxent.jar} and save it in the correct location.
The zoon function \code{GetMaxEnt} can orchestrate this for you.
Running MaxEnt also requires an up-to-date version of java
(which you may already have installed).
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
\code{\link{dismo::maxent}}

Other model: \code{\link{BiomodModel}}; \code{\link{GBM}};
  \code{\link{LogisticRegression}};
  \code{\link{MachineLearn}}; \code{\link{OptGRaF}};
  \code{\link{QuickGRaF}}; \code{\link{RandomForest}};
  \code{\link{mgcv}}
}

