% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module}
\alias{import}
\alias{module}
\alias{use}
\title{Define Modules in R}
\usage{
module(expr = { }, parent = if (interactive()) baseenv() else
  parent.frame())

import(from, ..., into = parent.env(parent.frame()))

use(module, where = parent.frame())
}
\arguments{
\item{expr}{an expression.}

\item{parent}{(environment) the root of the local search path.}

\item{from}{(character, or unquoten expression) a package name}

\item{...}{(character, or unquoted expression) names to import from package}

\item{into}{(environment) the environment for imports. Always first on the
search path.}

\item{module}{(character, list or environment) a module}

\item{where}{(environment)}
}
\description{
Use \code{module} to define self contained organizational units. Modules have
their own search path. \code{import} can be used to import single objects
from a package. \code{use} will attach collections or packages.
}

