% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-coercion.R
\name{as.module}
\alias{as.module}
\title{Coercion for Modules}
\usage{
as.module(filename, parent = baseenv(), ...)
}
\arguments{
\item{filename}{(character or connection) a file to be sourced as module}

\item{parent}{(environment) the root of the local search path.}

\item{...}{arguments passed to \link{parse}}
}
\description{
Interfaces to and from modules.
}

