% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NATrees.R
\name{NATrees}
\alias{NATrees}
\title{Occurrences for 21 tree species from North America}
\usage{
NATrees(dateRange = NULL, species = "acer_saccharum")
}
\arguments{
\item{dateRange}{Optional 2-item numeric vector giving the years to use; default is the full range of the data (1960-2013)}

\item{species}{Which species should be extracted, with format genus_species (all lowercase); allowable values are: abies_balsamea acer_rubrum acer_saccharum betula_alleghaniensis betula_papyrifera fagus_grandifolia fraxinus_americana fraxinus_nigra larix_laricina picea_glauca picea_mariana picea_rubens pinus_banksiana pinus_resinosa pinus_strobus populus_grandidentata populous_tremuloides quercus_macrocarpa quercus_rubra thuja_occidentalis tsuga_canadensis}
}
\description{
Retrieves occurrences for 21 trees from eastern North America. Data originally sourced from USGS FIA forest inventory database, collected in the QUICC-FOR project ($URL). This dataset was originally published in $REF. Extent for this dataset is c(-97,-67,25,50)
}
\details{

}
\section{Data type}{
 presence/absence
}

\section{Version}{
 0.5
}

\section{Date submitted}{
  2016-06-16
}

\seealso{
Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceDataFrame}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{Lorem_ipsum_UK}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{SugarMaple}},
  \code{\link{UKAnophelesPlumbeus}}
}
\author{
Matt Talluto, \email{mtalluto@gmail.com}
}
