% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCEP.R
\name{NCEP}
\alias{NCEP}
\title{Covariate module: NCEP}
\usage{
NCEP(extent = c(-5, 5, 50, 60), variables = "hgt", status.bar = FALSE)
}
\arguments{
\item{extent}{A numeric vector of length 4 giving the coordinates of the 
rectangular region within which to carry out the analysis, in the 
order: xmin, xmax, ymin, ymax.}

\item{variables}{Character vector of which variables to select.
See NCEP.gather for more options but basic options are
      *‘air’*    Air Temperature                 deg K    
      *‘hgt’*    Geopotential Height             m        
      *‘rhum’*   Relative Humidity               %        
      *‘shum’*   Specific Humidity               kg/kg    
      *‘omega’*  Omega [Vertical Velocity]       Pascal/s 
      *‘uwnd’*   U-Wind Component [East/West]    m/s      
      *‘vwnd’*   V-Wind Component [North/South]  m/s}

\item{status.bar}{Logical. Should a status bar be shown indicating the percentage of completion?}
}
\description{
Covariate module to grab coarse resolution environmental data from NCEP
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
\code{\link{RNCEP::NCEP.gather}}

Other covariate: \code{\link{AirNCEP}},
  \code{\link{Bioclim}}, \code{\link{LocalRaster}},
  \code{\link{UKAir}}, \code{\link{UKBioclim}}
}

