% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NoProcess.R
\name{NoProcess}
\alias{NoProcess}
\title{Process module: NoProcess}
\usage{
NoProcess(.data)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}
}
\value{
a Raster object (class from the raster package) with the gridded
     covariates used to train and predict from the SDM.
}
\description{
Process module that does nothing. A place holder for if nothing should be
done to the data before modelling.
}
\seealso{
Other process: \code{\link{BackgroundAndCrossvalid}};
  \code{\link{Background}}; \code{\link{BiasBackground}};
  \code{\link{Crossvalidate}};
  \code{\link{OneHundredBackground}};
  \code{\link{OneThousandBackground}};
  \code{\link{PartitionDisc}}
}

