% Generated by roxygen2 (4.0.2): do not edit by hand
\name{NoProcess}
\alias{NoProcess}
\title{Process module: NoProcess}
\value{
a Raster* object (class from the raster package) with the gridded
     covariates used to train and predict from the SDM.
}
\description{
Process module that does nothing. A place holder for if nothing should be
done to the data before modelling.
}

