% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/OneHundredBackground.R
\name{OneHundredBackground}
\alias{OneHundredBackground}
\title{Process module: OneHundredBackground}
\usage{
OneHundredBackground(.data)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}
}
\description{
Process module to generate up to 100 background records at random in
     cells of ras and return these along with the presence only data.
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other process: \code{\link{BackgroundAndCrossvalid}};
  \code{\link{Background}}; \code{\link{BiasBackground}};
  \code{\link{Crossvalidate}}; \code{\link{NoProcess}};
  \code{\link{OneThousandBackground}};
  \code{\link{PartitionDisc}}
}

