% Generated by roxygen2 (4.0.2): do not edit by hand
\name{OneHundredBackground}
\alias{OneHundredBackground}
\title{Process module: OneHundredBackground}
\usage{
OneHundredBackground(occurrence, ras)
}
\description{
Process module to generate up to 100 background records at random in
     cells of ras and return these along with the presence only data.
}

