% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneHundredBackground.R
\name{OneHundredBackground}
\alias{OneHundredBackground}
\title{Process module: OneHundredBackground}
\usage{
OneHundredBackground(.data, seed = NULL)
}
\arguments{
\item{.data}{\strong{Internal parameter, do not use in the workflow function}. \code{.data} is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. \code{.data} is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.}

\item{seed}{Numeric used with \code{\link[base]{set.seed}}}
}
\description{
Process module to generate up to 100 background records at random in
     cells of ras and return these along with the presence only data.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence-only
}

\seealso{
Other process: \code{\link{AddRandomUniformPredictors}},
  \code{\link{BackgroundAndCrossvalid}},
  \code{\link{Background}}, \code{\link{Bootstrap}},
  \code{\link{CarolinaWrenValidation}},
  \code{\link{Clean}}, \code{\link{Crossvalidate}},
  \code{\link{JitterOccurrence}},
  \code{\link{LonLatToCovariates}}, \code{\link{MESSMask}},
  \code{\link{NoProcess}},
  \code{\link{OneThousandBackground}},
  \code{\link{PartitionDisc}}, \code{\link{RemoveNAs}},
  \code{\link{StandardiseCov}},
  \code{\link{SubsampleOccurrence}},
  \code{\link{TargetGroupBackground}},
  \code{\link{Transform}}, \code{\link{addInteraction}},
  \code{\link{spThin}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
