% Generated by roxygen2 (4.0.1): do not edit by hand
\name{OneThousandBackground}
\alias{OneThousandBackground}
\title{Process module: OneThousandBackground}
\usage{
OneThousandBackground(occurrence, ras)
}
\description{
Process module to generate up to 1000 background records at random in
     cells of the covariate raster and return these along with the occurrence data.
}

