% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptGRaF.R
\name{OptGRaF}
\alias{OptGRaF}
\title{Model module: OptGRaF}
\usage{
OptGRaF(.df)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}. \code{.df} is data frame that combines the occurrence data and covariate data. \code{.df} is passed automatically in workflow from the process module(s) to the model module(s) and should not be passed by the user.}
}
\description{
Model module to fit a (slow) GRaF model, with parameter optimisation.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence/absence
}

\seealso{
Other model: \code{\link{BiomodModel}},
  \code{\link{Domain}}, \code{\link{GBM}},
  \code{\link{LogisticRegression}},
  \code{\link{MachineLearn}}, \code{\link{MaxEnt}},
  \code{\link{MaxLike}}, \code{\link{MaxNet}},
  \code{\link{MyMaxLike}}, \code{\link{NullModel}},
  \code{\link{QuickGRaF}}, \code{\link{RandomForest}},
  \code{\link{StochasticLogisticRegression}},
  \code{\link{mgcv}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
