% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceMeasures.R
\name{PerformanceMeasures}
\alias{PerformanceMeasures}
\title{Output module: PerformanceMeasures}
\usage{
PerformanceMeasures(.model, .ras, threshold = 0.5)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{threshold}{A chosen threshold value for measures that need 0/1 predictions}
}
\description{
Calculate a suite of performance metrics on either crossvalidation, external validation data or (at your own risk) in-sample validation.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other output: \code{\link{Figshare}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{PredictNewAreaMap}}, \code{\link{PrintMap}},
  \code{\link{ResponseCurve}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SurfaceMap}}
}

