% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerformanceMeasures.R
\name{PerformanceMeasures}
\alias{PerformanceMeasures}
\title{Output module: PerformanceMeasures}
\usage{
PerformanceMeasures(model, ras, threshold = 0.5)
}
\arguments{
\item{threshold}{A chosen threshold value for measures that need 0/1 predictions}
}
\description{
Calculate a suite of performance metrics on either crossvalidation or external validation data
}

