% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PredictNewAreaMap.R
\name{PredictNewAreaMap}
\alias{PredictNewAreaMap}
\title{Output module: PredictNewAreaMap}
\usage{
PredictNewAreaMap(model, ras, extent)
}
\arguments{
\item{extent}{Length 4 character vector giving the new extent to be predicted over. Must be within the covariate data collected in the covariate module.}
}
\description{
Output module. Predict a model across a new area.
}

