% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/PredictNewAreaMap.R
\name{PredictNewAreaMap}
\alias{PredictNewAreaMap}
\title{Output module: PredictNewAreaMap}
\usage{
PredictNewAreaMap(.model, .ras, extent = as.vector(.ras@extent))
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{extent}{Length 4 character vector giving the new extent to be predicted over. Must be within the covariate data collected in the covariate module.}
}
\description{
Output module. Predict a model across a new area.
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other output: \code{\link{Figshare}};
  \code{\link{InteractiveCovariateMap}};
  \code{\link{InteractiveMap}};
  \code{\link{PerformanceMeasures}};
  \code{\link{PrintMap}}; \code{\link{ResponseCurve}};
  \code{\link{SameTimePlaceMap}}; \code{\link{SurfaceMap}}
}

