% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/PrintMap.R
\name{PrintMap}
\alias{PrintMap}
\title{Output module: PrintMap}
\usage{
PrintMap(.model, .ras)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}
}
\description{
Plot a map of predicted surface.
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other output: \code{\link{Figshare}};
  \code{\link{InteractiveCovariateMap}};
  \code{\link{InteractiveMap}};
  \code{\link{PerformanceMeasures}};
  \code{\link{PredictNewAreaMap}};
  \code{\link{ResponseCurve}};
  \code{\link{SameTimePlaceMap}}; \code{\link{SurfaceMap}}
}

