% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickGRaF.R
\name{QuickGRaF}
\alias{QuickGRaF}
\title{Model module: QuickGRaF}
\usage{
QuickGRaF(.df, l = NULL)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}. \code{.df} is data frame that combines the occurrence data and covariate data. \code{.df} is passed automatically in workflow from the process module(s) to the model module(s) and should not be passed by the user.}

\item{l}{lengthscale parameter, a positive number controlling the complexity
of the fitted model with respect to each covariate.
Can either be a single number (in which case it is used for all covariates),
a vector of numbers of the same length as the number of covariates,
or \code{NULL} in which case GRaF will calculate a fixed approximation to
a sensible lengthscale.}
}
\description{
Model module to fit a quick GRaF model, without parameter optimisation.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other model: \code{\link{BiomodModel}}, \code{\link{GBM}},
  \code{\link{LogisticRegression}},
  \code{\link{MachineLearn}}, \code{\link{MaxEnt}},
  \code{\link{OptGRaF}}, \code{\link{RandomForest}},
  \code{\link{mgcv}}
}

