% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{RandomForest}
\alias{RandomForest}
\title{Model module: RandomForest}
\usage{
RandomForest(.df)
}
\arguments{
\item{.df}{\strong{Internal parameter, do not use in the workflow function}. \code{.df} is data frame that combines the occurrence data and covariate data. \code{.df} is passed automatically in workflow from the process module(s) to the model module(s) and should not be passed by the user.}
}
\value{
A model object with a valid predict method
}
\description{
Model module to fit a simple RandomForest classification model
}
\section{Version}{
 1.1
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence/absence, presence/background
}

\seealso{
Other model: \code{\link{BiomodModel}},
  \code{\link{Domain}}, \code{\link{GBM}},
  \code{\link{LogisticRegression}},
  \code{\link{MachineLearn}}, \code{\link{MaxEnt}},
  \code{\link{MaxLike}}, \code{\link{MaxNet}},
  \code{\link{MyMaxLike}}, \code{\link{OptGRaF}},
  \code{\link{QuickGRaF}},
  \code{\link{StochasticLogisticRegression}},
  \code{\link{mgcv}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
