% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResponseCurve.R
\name{ResponseCurve}
\alias{ResponseCurve}
\title{Output module: ResponseCurve}
\usage{
ResponseCurve(model, ras, cov = 1)
}
\arguments{
\item{model}{}

\item{ras}{}

\item{cov}{which of the covariates to plot the response against}
}
\description{
Plot a detailed conditional response curve from the model
against one covariate.
}

