% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseCurve.R
\name{ResponseCurve}
\alias{ResponseCurve}
\title{Output module: ResponseCurve}
\usage{
ResponseCurve(.model, .ras, cov = NULL)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{cov}{which of the covariates to plot the response against.
If \code{NULL}, the function is executed for all covariates, one after another.
Else it should be a single numeric identifying the covariate to plot.}
}
\description{
Plot a detailed conditional response curve from the model
against one covariate.
}
\section{Version}{
 1.1
}

\section{Date submitted}{
 2017-04-11
}

\seealso{
Other output: \code{\link{Appify}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}}, \code{\link{NoOutput}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurveViz}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
