% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SameTimePlaceMap.R
\name{SameTimePlaceMap}
\alias{SameTimePlaceMap}
\title{Output module: SameTimePlaceMap}
\usage{
SameTimePlaceMap(.model, .ras)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}
}
\value{
A Raster object giving the probabilistic model predictions for each
     cell of ras
}
\description{
Output module. A function for outputing the raster of the predictions from an analysis
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\seealso{
Other output: \code{\link{AIC}}, \code{\link{AUC}},
  \code{\link{Appify}}, \code{\link{CoefficientPlot}},
  \code{\link{CovHistograms}}, \code{\link{DataSummary}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{InteractiveOccurrenceMap}},
  \code{\link{NoOutput}}, \code{\link{PairPlot}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ReliabilityPlot}},
  \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}}, \code{\link{ResponsePlot}},
  \code{\link{SeparatePA}}, \code{\link{SurfaceMap}},
  \code{\link{VariableImportance}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
