% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SameTimePlaceMap}
\alias{SameTimePlaceMap}
\title{Output module: SameTimePlaceMap}
\usage{
SameTimePlaceMap(model, ras)
}
\value{
A Raster object giving the probabilistic model predictions for each
     cell of ras
}
\description{
Output module. A function for outputing the raster of the predictions from an analysis
}

