% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpOcc.R
\name{SpOcc}
\alias{SpOcc}
\title{Occurrence module: SpOcc}
\usage{
SpOcc(species, extent, databases = "gbif")
}
\arguments{
\item{species}{A character string giving the species name.
 e.g. 'Anopheles plumbeus'.}

\item{extent}{A numeric vector of length 4 giving the coordinates of the
 rectangular region within which to carry out the analysis, in the
 order: xmin, xmax, ymin, ymax.}

\item{databases}{A character vector giving the databases to use.
Choose from "gbif", "bison", "inat", "ebird", "ecoengine", "antweb".
Defaults to gbif. NB I have had some problems with databases other
than gbif.}
}
\description{
Occurrence module to collect occurrence data from a number of data bases.
}
\seealso{
\code{\link{spocc::occ}}
}

