% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpOcc.R
\name{SpOcc}
\alias{SpOcc}
\title{Occurrence module: SpOcc}
\usage{
SpOcc(species = "Myotis daubentonii", extent = c(-1, 0, 51, 52),
  databases = "gbif", type = "presence", limit = 10000)
}
\arguments{
\item{species}{A character string giving the species name.
e.g. 'Anopheles plumbeus'.}

\item{extent}{A numeric vector of length 4 giving the coordinates of the 
rectangular region within which to carry out the analysis, in the 
order: xmin, xmax, ymin, ymax.}

\item{databases}{A character vector giving the databases to use.
Choose from "gbif", "bison", "inat", "ebird", "ecoengine", "antweb".
Defaults to gbif. NB I have had some problems with databases other
than gbif.}

\item{type}{A string giving the type of occurrence data to treat these records as.
Must be one of \code{'presence'} (the default and most common use case),
\code{'absence'} or \code{'background'} (which may be used to implement
a 'target background group' approach for overcoming observation bias).}

\item{limit}{The maximum number of occurrence records to return.}
}
\description{
Occurrence module to collect occurrence data from a number of data bases.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence-only, presence/absence, abundance, proportion
}

\seealso{
\code{\link{spocc::occ}}

Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceDataFrame}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{Lorem_ipsum_UK}}, \code{\link{NATrees}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}},
  \code{\link{SugarMaple}},
  \code{\link{UKAnophelesPlumbeus}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\concept{occurrence}
