% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SurfaceMap}
\alias{SurfaceMap}
\title{Output module: SurfaceMap}
\usage{
SurfaceMap(model, ras, dir = ".")
}
\arguments{
\item{dir}{Where to save figures. Defaults to the working directory.}
}
\description{
Make a png of a map of predicted surface.
}

