% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurfaceMap.R
\name{SurfaceMap}
\alias{SurfaceMap}
\title{Output module: SurfaceMap}
\usage{
SurfaceMap(.model, .ras, dir = NULL)
}
\arguments{
\item{.model}{\strong{Internal parameter, do not use in the workflow function}. \code{.model} is list of a data frame (\code{data}) and a model object (\code{model}). \code{.model} is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.}

\item{.ras}{\strong{Internal parameter, do not use in the workflow function}. \code{.ras} is a raster layer, brick or stack object. \code{.ras} is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.}

\item{dir}{Directory where plots are saved. If NULL (default) then plots are not saved.}
}
\description{
Make a png of a map of predicted surface.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\seealso{
Other output: \code{\link{Appify}},
  \code{\link{GenerateCovariateReport}},
  \code{\link{InteractiveCovariateMap}},
  \code{\link{InteractiveMap}},
  \code{\link{InteractiveOccurrenceMap}},
  \code{\link{NoOutput}},
  \code{\link{PerformanceMeasures}},
  \code{\link{PredictNewRasterMap}},
  \code{\link{PrintMap}}, \code{\link{PrintOccurrenceMap}},
  \code{\link{ROCcurve}}, \code{\link{ResponseCurveViz}},
  \code{\link{ResponseCurve}},
  \code{\link{SameTimePlaceMap}}, \code{\link{SeparatePA}},
  \code{\link{VariableImportance}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
