% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UKAir.R
\name{UKAir}
\alias{UKAir}
\title{Covariate module: UKAir}
\usage{
UKAir()
}
\description{
Return some UK air temperature data from NCEP. Data is bundled with package and has the extent \code{c(-10, 10, 45, 65)}.
}
\keyword{Covariate}

