% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKAir.R
\name{UKAir}
\alias{UKAir}
\title{Covariate module: UKAir}
\usage{
UKAir()
}
\description{
Return some UK air temperature data from NCEP. Data is bundled with package and has the extent \code{c(-10, 10, 45, 65)}.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\seealso{
Other covariate: \code{\link{AirNCEP}},
  \code{\link{Bioclim_future}}, \code{\link{Bioclim}},
  \code{\link{CarolinaWrenRasters}},
  \code{\link{LocalRaster}}, \code{\link{NCEP}},
  \code{\link{NaiveRandomRaster}}, \code{\link{UKBioclim}},
  \code{\link{malariaAtlas_covariates}}
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
