% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKAnophelesPlumbeus.R
\name{UKAnophelesPlumbeus}
\alias{UKAnophelesPlumbeus}
\title{Occurrence module: UK Anopheles plumbeus data}
\usage{
UKAnophelesPlumbeus()
}
\description{
Return some Anopheles plumbeus data that is bundled with package.
 Data taken from GBIF and has the extent \code{c(-10, 10, 45, 65)}.
}
\section{Version}{
 1.0
}

\section{Date submitted}{
 2015-11-13
}

\section{Data type}{
 presence-only
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other occurrence: \code{\link{AnophelesPlumbeus}},
  \code{\link{CWBZimbabwe}}, \code{\link{CarolinaWrenPA}},
  \code{\link{CarolinaWrenPO}},
  \code{\link{LocalOccurrenceData}},
  \code{\link{Lorem_ipsum_UK}}, \code{\link{NATrees}},
  \code{\link{NBNdataByName}},
  \code{\link{NaiveRandomPresenceAbsence}},
  \code{\link{NaiveRandomPresence}}, \code{\link{SpOcc}},
  \code{\link{SugarMaple}}
}

