% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UKAnophelesPlumbeus.R
\name{UKAnophelesPlumbeus}
\alias{UKAnophelesPlumbeus}
\title{Return some Anopheles plumbeus data that is bundled with package.
 Data taken from GBIF and has the extent \code{c(-10, 10, 45, 65)}.}
\usage{
UKAnophelesPlumbeus()
}
\description{
Return some Anopheles plumbeus data that is bundled with package.
 Data taken from GBIF and has the extent \code{c(-10, 10, 45, 65)}.
}

