% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKBioclim.R
\name{UKBioclim}
\alias{UKBioclim}
\title{Covariate module: UKBioclim}
\usage{
UKBioclim()
}
\description{
Load Bioclim rasters at 5 degree resolution for the UK
	(extent \code{c(-10, 10, 45, 65)}).
}
\details{
The first time this module is used the data are downloaded from the web,
	subsequently a local copy is used so it's much faster.
	This is essentially a wrapper around the `raster` function `getData`,
	see `?getData' for more details of what that function does.
Use the \code{Bioclim} module for bioclim data with other extents.
}
\author{
ZOON Developers, \email{zoonproject@gmail.com}
}
\seealso{
Other covariate: \code{\link{AirNCEP}},
  \code{\link{Bioclim}}, \code{\link{LocalRaster}},
  \code{\link{NCEP}}, \code{\link{UKAir}}
}

