% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSearchPath.R
\name{getSearchPath}
\alias{getSearchPath}
\alias{getSearchPathNames}
\title{Get the search path of an environment}
\usage{
getSearchPath(where = parent.frame())

getSearchPathNames(where = parent.frame())
}
\arguments{
\item{where}{(environment)}
}
\description{
Returns a list with the environments or names of the environments on the
search path. These functions are used for testing, use \link{search} instead.
}

